% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dataset.R
\name{add_dataset}
\alias{add_dataset}
\title{Adds new datasets to list}
\usage{
add_dataset(dataset, return = T)
}
\arguments{
\item{dataset}{dataset to add}

\item{return}{Should TRUE be returned if dataset is added successfully}
}
\value{

}
\description{
Adds new datasets to list
}
