% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{calculates adjacency matrix from similarity matrix}
\usage{
adjacency(data, method, alpha, theta, beta)
}
\arguments{
\item{data}{fimilarity matrix}

\item{method}{adjacency function}

\item{alpha}{alpha parameter for sigmoid function}

\item{theta}{theta parameter for sigmoid function}

\item{beta}{beta parameter for power function}
}
\value{

}
\description{
calculates adjacency matrix from similarity matrix
}
