% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_.R
\name{adjacency_}
\alias{adjacency_}
\title{Adds adjacency to cor_list}
\usage{
adjacency_(cor_list, method = "power", alpha = 2, theta = 0.5, beta = 1)
}
\arguments{
\item{cor_list}{cor_list object}

\item{method}{adjacency method}

\item{alpha}{shape of sigmoid function}

\item{theta}{turning point of sigmoid function}

\item{beta}{exponent of power function}
}
\value{

}
\description{
Adds adjacency to cor_list
}
