% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anal_data_grouped_var.R
\name{anal_data_grouped_var}
\alias{anal_data_grouped_var}
\title{Evalueates data groupwise and produces venn diagrams and table}
\usage{
anal_data_grouped_var(
  variables,
  observations,
  groups,
  expr1,
  expr2,
  name,
  observations.set,
  data.name,
  type,
  dataset,
  save.separete = T,
  save.variable.groups = T,
  show = T,
  export.plot = T,
  export.table = F,
  return.export.table = F
)
}
\arguments{
\item{variables}{variables}

\item{observations}{observations}

\item{groups}{group definition}

\item{expr1}{function(x) how columns should be evaluated}

\item{expr2}{function(x) how groups should be evaluated}

\item{name}{name}

\item{observations.set}{set of observations}

\item{data.name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{save.separete}{save results from group evaluation in variables data}

\item{save.variable.groups}{save variable groups in variables data}

\item{show}{show plot}

\item{export.plot}{exports plot to pdf}

\item{export.table}{exports table of proteins}

\item{return.export.table}{return table of proteins}
}
\value{

}
\description{
Evalueates data groupwise and produces venn diagrams and table
}
