% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anal_expression.R
\name{anal_expression}
\alias{anal_expression}
\title{Calculates Fold change and p-value from given dataframe or defined variables, observations and groups}
\usage{
anal_expression(
  data,
  variables = "default",
  observations = "default",
  groups,
  control,
  data.name,
  var.equal = T,
  type = "LFQ",
  observations.set,
  dataset,
  plot = T,
  view = F,
  save = F,
  return = T
)
}
\arguments{
\item{data}{(optional) data frame}

\item{variables}{(optional) variables definition}

\item{observations}{(optional) observatoins definition}

\item{groups}{groups}

\item{control}{(optional) define control sample (case / control)}

\item{data.name}{(optional)}

\item{var.equal}{t-test with equal variance}

\item{type}{data type}

\item{observations.set}{observations.set}

\item{dataset}{dataset}

\item{plot}{Plot?}

\item{view}{View results matrix?}

\item{save}{Save results matrix?}

\item{return}{Return results matrix?}
}
\value{

}
\description{
Calculates Fold change and p-value from given dataframe or defined variables, observations and groups
}
