% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_name.R
\name{ask_name}
\alias{ask_name}
\title{Ask user for name of no name given}
\usage{
ask_name(name, message)
}
\arguments{
\item{name}{name}

\item{message}{message}
}
\value{

}
\description{
Ask user for name of no name given
}
