% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_limit_breaks.R
\name{axis_limit_breaks}
\alias{axis_limit_breaks}
\title{Helps make nice axis limit breaks}
\usage{
axis_limit_breaks(values, break.space = 2)
}
\arguments{
\item{values}{data values}

\item{break.space}{space between breaks}
}
\value{

}
\description{
Helps make nice axis limit breaks
}
