% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cor_list.R
\name{build_cor_list}
\alias{build_cor_list}
\title{Builds cor_list object with default parameters}
\usage{
build_cor_list(
  data,
  min.0 = 0.5,
  shift = 1.8,
  width = 0.3,
  seed.impute = 123,
  similarity.method = "preserve",
  adjacency.method = "none"
)
}
\arguments{
\item{data}{data}

\item{min.0}{threshold to keep proteins}

\item{shift}{shift in standard deviations}

\item{width}{width in standard deviations}

\item{seed.impute}{seed}

\item{similarity.method}{method to calculate similarity matrix}

\item{adjacency.method}{method to calculate adjacency matrix}
}
\value{

}
\description{
Builds cor_list object with default parameters
}
