% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_database.R
\name{check_database}
\alias{check_database}
\title{Checks if database is available}
\usage{
check_database(database = "UniProt", taxId)
}
\arguments{
\item{database}{annotation database}

\item{taxId}{taxonomy Id}
}
\value{

}
\description{
Checks if database is available
}
