% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_GO_term.R
\name{choose_GO_terms}
\alias{choose_GO_terms}
\title{Interface for user to define a list of GO terms}
\usage{
choose_GO_terms(return.ID = F, graphics = T)
}
\arguments{
\item{return.ID}{return GO ID or terms}

\item{graphics}{Choose from graphic interface}
}
\value{

}
\description{
Interface for user to define a list of GO terms
}
