% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_2w_anova_i.R
\name{classify_2w_anova_i}
\alias{classify_2w_anova_i}
\title{Classifies protein changes from two-way ANOVA with interaction}
\usage{
classify_2w_anova_i(data, p.value.threshold = 0.05, fc.threshold = 0.5)
}
\arguments{
\item{data}{data frame output of two-way ANOVA with interactions}

\item{p.value.threshold}{threshold for p-values}

\item{fc.threshold}{absolute threshold for log2 fold-change}
}
\value{

}
\description{
Classifies protein changes from two-way ANOVA with interaction
}
