% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_protein_identification_3.R
\name{classify_protein_identification_3}
\alias{classify_protein_identification_3}
\title{Classifies proteins based on identification threshold}
\usage{
classify_protein_identification_3(
  data.eval,
  lower.limit = 0,
  dataset,
  plot = T,
  save = F
)
}
\arguments{
\item{data.eval}{data from eval_ function}

\item{lower.limit}{values above limit are considered positive/identified}

\item{dataset}{dataset}

\item{plot}{Plot venn diagram?}

\item{save}{Save results list?}
}
\value{

}
\description{
Classifies proteins based on identification threshold
}
