% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_protein_identification_n.R
\name{classify_protein_identification_n}
\alias{classify_protein_identification_n}
\title{Classifies proteins based on identification threshold}
\usage{
classify_protein_identification_n(data.eval, lower.limit = 0, plot = T)
}
\arguments{
\item{data.eval}{data from eval_ function}

\item{lower.limit}{values above limit are considered positive/identified}

\item{plot}{Plot venn diagram?}

\item{save}{Save results list?}
}
\value{

}
\description{
Classifies proteins based on identification threshold
}
