% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_proteins.R
\name{cleanup_proteins}
\alias{cleanup_proteins}
\title{Removes proteins based on prefixes}
\usage{
cleanup_proteins(data, prefix = c("CON_", "REV_"))
}
\arguments{
\item{data}{data}

\item{prefix}{prefix to be removed}
}
\value{

}
\description{
Removes proteins based on prefixes
}
