% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_proteins_m.R
\name{cleanup_proteins_m}
\alias{cleanup_proteins_m}
\title{Removes proteins based on prefixes from columns}
\usage{
cleanup_proteins_m(matrix, prefix = c("CON_", "REV_"))
}
\arguments{
\item{matrix}{data}

\item{prefix}{prefix to be removed}
}
\value{

}
\description{
Removes proteins based on prefixes from columns
}
