% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_proteins_v.R
\name{cleanup_proteins_v}
\alias{cleanup_proteins_v}
\title{Removes proteins based on prefixes from vector}
\usage{
cleanup_proteins_v(vector, prefix = c("CON_", "REV_", "NA_"))
}
\arguments{
\item{vector}{vector}

\item{prefix}{prefix to be removed}
}
\value{

}
\description{
Removes proteins based on prefixes from vector
}
