% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_groups.R
\name{collapse_groups}
\alias{collapse_groups}
\title{Collapses groups of data frame by specified function}
\usage{
collapse_groups(data, FUN = mean, group.column = "groups")
}
\arguments{
\item{data}{tibble data frame}

\item{FUN}{function to combine groups}

\item{group.column}{column specifying the groups}
}
\value{

}
\description{
Collapses groups of data frame by specified function
}
