% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_groups_.R
\name{collapse_groups_}
\alias{collapse_groups_}
\title{Collapses groups of data frame by specified function and adds as list entry}
\usage{
collapse_groups_(
  analysis_list,
  FUN = mean,
  group.column = "groups",
  data.name = "raw_data",
  new.name = "grouped_data"
)
}
\arguments{
\item{analysis_list}{analysis_list}

\item{FUN}{function to combine groups}

\item{group.column}{column specifying the groups}

\item{data.name}{name of data frame entry}

\item{new.name}{new entry name}
}
\value{

}
\description{
Collapses groups of data frame by specified function and adds as list entry
}
