% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_rows.R
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapses rows of data frame by specified function and adds as list entry}
\usage{
collapse_rows(data_, FUN = mean, by = "All", input, output)
}
\arguments{
\item{data_}{data}

\item{FUN}{function to combine groups}

\item{by}{column specifying the groups}

\item{input}{name of input data frame}

\item{new.name}{name of output data frame (defaults to input data name)}
}
\value{

}
\description{
Collapses rows of data frame by specified function and adds as list entry
}
