% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_groups.R
\name{combine_data_groups}
\alias{combine_data_groups}
\title{Sums up numerical and logical data groupwise}
\usage{
combine_data_groups(data, groups, method)
}
\arguments{
\item{data}{data}

\item{groups}{groups}

\item{method}{method to use for combination}
}
\value{

}
\description{
Sums up numerical and logical data groupwise
}
