% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_.R
\name{connectivity_}
\alias{connectivity_}
\title{Add connectivity to cor_list}
\usage{
connectivity_(cor_list, scale = T, plot = T, fun.enrich)
}
\arguments{
\item{cor_list}{cor_list object}

\item{scale}{Scale connectivity values}

\item{plot}{plot results}

\item{fun.enrich}{do functional enrichment on connectivity values}
}
\value{

}
\description{
Add connectivity to cor_list
}
