% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_.R
\name{cor_}
\alias{cor_}
\title{Calculates correlation matrix and return cor_list}
\usage{
cor_(data, method = "pearson", set.diag.0 = F)
}
\arguments{
\item{data}{LFQ data}

\item{method}{method to calculate correlation}

\item{set.diag.0}{set diagonal to zero}
}
\value{

}
\description{
Calculates correlation matrix and return cor_list
}
