% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Computes a correlation matrix}
\usage{
cor_matrix(
  variables,
  observations,
  observations.set,
  data.name,
  dataset,
  similarity.method = "none",
  adjacency.method = "none",
  alpha,
  theta,
  beta,
  clustering.method = "average"
)
}
\arguments{
\item{variables}{variables (see get variables)}

\item{observations}{observations (see get observations)}

\item{observations.set}{set of observations}

\item{data.name}{data name}

\item{dataset}{dataset}

\item{similarity.method}{similarity function}

\item{adjacency.method}{adjacency function}

\item{alpha}{alpha parameter for sigmoid function}

\item{theta}{theta parameter for sigmoid function}

\item{beta}{beta parameter for power function}

\item{clustering.method}{method for function hclust}
}
\value{

}
\description{
Computes a correlation matrix
}
