% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame2tibble.R
\name{data.frame2tibble}
\alias{data.frame2tibble}
\title{Transforms data frames to tibble and adds column for row names}
\usage{
data.frame2tibble(data, row.names)
}
\arguments{
\item{data}{data frame with row names}

\item{row.names}{name for row names vector}
}
\value{

}
\description{
Transforms data frames to tibble and adds column for row names
}
