% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2tibble.R
\name{data2tibble}
\alias{data2tibble}
\title{Transforms any data type to a tibble}
\usage{
data2tibble(data, row.names)
}
\arguments{
\item{data}{supported data types: matrix, data.frame}

\item{row.names}{name for row names vector}
}
\value{

}
\description{
Transforms any data type to a tibble
}
