% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2vector.R
\name{data2vector}
\alias{data2vector}
\title{Returns names vector from data frame}
\usage{
data2vector(data, values, names)
}
\arguments{
\item{data}{data frame, tibble or matrix}

\item{values}{column for vector values; by default second column}

\item{names}{column for vector names; by default first column}
}
\value{

}
\description{
Returns names vector from data frame
}
