% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2venn.R
\name{data2venn}
\alias{data2venn}
\title{Evalueates data groupwise and produces venn diagrams and table}
\usage{
data2venn(
  variables,
  observations,
  groups,
  expr1,
  expr2,
  name,
  observations.set,
  data.name,
  type,
  dataset,
  save.separete = T,
  save = T,
  export.table = F,
  return.export.table = F
)
}
\arguments{
\item{variables}{variables}

\item{observations}{observations}

\item{groups}{group definition}

\item{expr1}{function(x) how columns should be evaluated}

\item{expr2}{function(x) how groups should be evaluated}

\item{name}{name}

\item{observations.set}{set of observations}

\item{data.name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{save.separete}{save results from group evaluation}

\item{save}{save venn diagram groups}

\item{export.table}{export table}

\item{return.export.table}{return exported table}
}
\value{

}
\description{
Evalueates data groupwise and produces venn diagrams and table
}
