% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame2tibble.R
\name{data_frame2tibble}
\alias{data_frame2tibble}
\title{Transforms data frames to tibble and adds column for row names}
\usage{
data_frame2tibble(data.frame, row.names = "observations")
}
\arguments{
\item{data.frame}{data frame with row names}

\item{row.names}{name for row names vector}
}
\value{

}
\description{
Transforms data frames to tibble and adds column for row names
}
