% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_helpers.R
\name{data_input}
\alias{data_input}
\title{Handles list input for functions}
\usage{
data_input(data_, input)
}
\arguments{
\item{data_}{data_}

\item{input}{name of input data}
}
\value{

}
\description{
Handles list input for functions
}
