% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_category.R
\name{define.category}
\alias{define.category}
\title{Defines new observation}
\usage{
define.category(observations)
}
\arguments{
\item{observations}{observations data frame}
}
\value{

}
\description{
Defines new observation
}
