% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram2table_.R
\name{dendrogram2table_}
\alias{dendrogram2table_}
\title{Adds table representation of dendrogram}
\usage{
dendrogram2table_(data_, dendrogram = "dend_x")
}
\arguments{
\item{data_}{data list}
}
\value{

}
\description{
Adds table representation of dendrogram
}
