% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram_.R
\name{dendrogram_}
\alias{dendrogram_}
\title{Computes a dendrogram from adjacency matrix in cor_object}
\usage{
dendrogram_(cor_list, method = "average")
}
\arguments{
\item{cor_list}{cor_list object}

\item{method}{clustering method}
}
\value{

}
\description{
Computes a dendrogram from adjacency matrix in cor_object
}
