% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2groupedList.R
\name{df2groupedList}
\alias{df2groupedList}
\title{Stores data frame in a list and groups adjacent columns}
\usage{
df2groupedList(x, min.similarity = 8, min.groupsize = 6)
}
\arguments{
\item{x}{data frame}

\item{min.similarity}{minimal similarity of column names}

\item{min.groupsize}{minimal group size}
}
\value{
grouped list of data frame columns
}
\description{
Stores data frame in a list and groups adjacent columns
}
