% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_1w_anova.R
\name{do_1w_anova}
\alias{do_1w_anova}
\title{Title}
\usage{
do_1w_anova(
  analysis_list,
  formula = "log2(x) ~ groups",
  data.name = "raw_data"
)
}
\arguments{
\item{data.name}{}
}
\value{

}
\description{
Title
}
