% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_2w_anova_i.R
\name{do_2w_anova_i}
\alias{do_2w_anova_i}
\title{Performs a two-way ANOVA with an interaction term between variables}
\usage{
do_2w_anova_i(
  analysis_list,
  formula = "log2(x) ~ genotype * treatment",
  data.name = "raw_data",
  balanced = T
)
}
\arguments{
\item{analysis_list}{list containing data}

\item{formula}{formula for model fitting (see aov() function)}

\item{data.name}{name of data in list}

\item{balanced}{Are analysis groups balanced (same number each)?}
}
\value{

}
\description{
Performs a two-way ANOVA with an interaction term between variables
}
