% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_abundance_summary.R
\name{do_abundance_summary}
\alias{do_abundance_summary}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_abundance_summary(
  data_,
  xlab = "Protein rank",
  ylab = "median log10 LFQ",
  point.size = 1,
  view = T,
  input,
  output.data = "data_abundance",
  output.plot = "plot_abundance"
)
}
\arguments{
\item{data_}{list or tibble}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{point.size}{size of data points}

\item{view}{view plot}

\item{input}{name of input data}

\item{output.data}{name for output data}

\item{output.plot}{name for output plot data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
