% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_analysis.R
\name{do_analysis}
\alias{do_analysis}
\title{Performs analysis}
\usage{
do_analysis(data_, data.name = "raw_data", plot = T)
}
\arguments{
\item{data_}{data list}

\item{data.name}{name of data to use}

\item{plot}{results}
}
\value{

}
\description{
Performs analysis
}
