% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_annotation_composition.R
\name{do_annotation_composition}
\alias{do_annotation_composition}
\title{Calculates composition of given annotations}
\usage{
do_annotation_composition(
  list_,
  TERM2GENE,
  correct.fractions.by,
  output = "data_annotation_composition"
)
}
\arguments{
\item{list_}{list containing protein sets}

\item{TERM2GENE}{TERM2GENE data frame}

\item{correct.fractions.by}{devide fractions by total annotated proteins}
}
\value{

}
\description{
Calculates composition of given annotations
}
