% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_anova.R
\name{do_anova}
\alias{do_anova}
\title{Performs one-way ANOVA}
\usage{
do_anova(analysis_list, formula = "log2(x) ~ groups", data.name = "raw_data")
}
\arguments{
\item{analysis_list}{list containing raw_data}

\item{formula}{formula given as string}

\item{data.name}{data to use for analysis}
}
\description{
Performs one-way ANOVA
}
