% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cormat.R
\name{do_cormat}
\alias{do_cormat}
\title{Do a correlation matrix analysis as described by Horvath et al.}
\usage{
do_cormat(
  data_,
  cor.method = "pearson",
  diag0 = T,
  similarity.method = "none",
  adjacency.method = "none",
  alpha,
  theta,
  beta,
  distance.method = "euclidean",
  clustering.method = "complete",
  dataset,
  plot = T,
  input,
  output = "data_cormat"
)
}
\arguments{
\item{data_}{list or tibble}

\item{cor.method}{function to compute correlation coefficients}

\item{diag0}{set autocorrelation to 0}

\item{similarity.method}{similarity function ("absolute", "preserve", "none")}

\item{adjacency.method}{adjacency function ("sigmoid", "power", "none")}

\item{alpha}{alpha parameter for sigmoid function}

\item{theta}{theta parameter for sigmoid function}

\item{beta}{beta parameter for power function}

\item{clustering.method}{method to cluster columns and rows ("complete",
"average"); see ?hclust for options}

\item{plot}{generate default plot with plot_cor_matrix}

\item{input}{name of input data}

\item{output}{name of output data}

\item{distance.m}{ethod method to calculate distance; see ?dist for options}
}
\description{
Do a correlation matrix analysis as described by Horvath et al.
}
