% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_dend_enrich.R
\name{do_dend_enrich}
\alias{do_dend_enrich}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_dend_enrich(
  data_,
  dend = "dend_y",
  ...,
  dataset,
  input = "data_cormat",
  output
)
}
\arguments{
\item{data_}{list or tibble}

\item{dend}{which dendrogram to use for analysis}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
