% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_enrichment_fisher.R
\name{do_enrichment_fisher}
\alias{do_enrichment_fisher}
\title{Performs overexpression enrichment on logical protein vector with specified Ontology}
\usage{
do_enrichment_fisher(proteins, database)
}
\arguments{
\item{proteins}{character vector}

\item{database}{database to use}
}
\value{

}
\description{
Performs overexpression enrichment on logical protein vector with specified Ontology
}
