% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_enrichment_ks.R
\name{do_enrichment_ks}
\alias{do_enrichment_ks}
\title{Performs gene set enrichment on numeric protein vector with specified Ontology}
\usage{
do_enrichment_ks(proteins, database)
}
\arguments{
\item{proteins}{numeric score vector}

\item{database}{database to use}
}
\value{

}
\description{
Performs gene set enrichment on numeric protein vector with specified Ontology
}
