% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_enrichment_mfisher.R
\name{do_enrichment_mfisher}
\alias{do_enrichment_mfisher}
\title{Performs overexpression enrichment on protein group vectors}
\usage{
do_enrichment_mfisher(proteins, database)
}
\arguments{
\item{proteins}{character vector}

\item{database}{database to use}
}
\value{

}
\description{
Performs overexpression enrichment on protein group vectors
}
