% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fun_enrich_dend.R
\name{do_fun_enrich_dend}
\alias{do_fun_enrich_dend}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_fun_enrich_dend(
  data_,
  dend = "dend_y",
  database,
  max.k = 10,
  ...,
  dataset,
  input,
  output
)
}
\arguments{
\item{data_}{list or tibble}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
