% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GSEA.R
\name{do_GSEA}
\alias{do_GSEA}
\title{Performs gene set enrichment on numeric protein vector with specified Ontology}
\usage{
do_GSEA(
  proteins,
  database,
  inverse = F,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 120,
  dataset,
  return.all = F,
  add.info = F
)
}
\arguments{
\item{proteins}{numeric score vector}

\item{database}{database to use}

\item{inverse}{enrich for terms in higher values/scores}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction), "hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}

\item{dataset}{dataset}

\item{return.all}{Should the enrichrObject be returned}

\item{add.info}{Add additional information (takes longer)}
}
\value{

}
\description{
Performs gene set enrichment on numeric protein vector with specified Ontology
}
