% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GSEA_GO.R
\name{do_GSEA_GO}
\alias{do_GSEA_GO}
\title{Performs gene set enrichment on numeric protein vector with specified Ontology}
\usage{
do_GSEA_GO(
  proteins,
  inverse = F,
  ontology = "CC",
  algorithm = "classic",
  threshold = 0.05,
  add.info = F
)
}
\arguments{
\item{proteins}{numeric score vector}

\item{inverse}{enrich for terms in higher values/scores}

\item{ontology}{GO ontology}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{threshold}{p-value/confidence threshold to exclude terms}

\item{add.info}{Add additional information (takes longer)}
}
\value{

}
\description{
Performs gene set enrichment on numeric protein vector with specified Ontology
}
