% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GSEA_Reactome.R
\name{do_GSEA_Reactome}
\alias{do_GSEA_Reactome}
\title{Performs gene set enrichment analysis using Reactome annotations}
\usage{
do_GSEA_Reactome(
  proteins,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  view = T,
  return.all = F,
  add.info = F
)
}
\arguments{
\item{proteins}{numeric vector of proteins scores}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none","hochberg" (Benjamini-Hochberg correction), "bonferroni", "holm", "hommel", "BH", "BY", "fdr"}

\item{view}{view results}

\item{return.all}{return enrichResult object; useful for further analysis of enrichment results}

\item{add.info}{add additional information to the results data frame}
}
\value{

}
\description{
Performs gene set enrichment analysis using Reactome annotations
}
