% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GSEA_topGO.R
\name{do_GSEA_topGO}
\alias{do_GSEA_topGO}
\title{Performs gene set enrichment on numeric protein vector with specified Ontology}
\usage{
do_GSEA_topGO(
  proteins,
  ontology = "CC",
  algorithm = "classic",
  threshold = 0.05,
  add.info = F
)
}
\arguments{
\item{proteins}{numeric score vector}

\item{ontology}{GO ontology}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{add.info}{Add additional information (takes longer)}

\item{pvalueCutoff}{p-value cutoff for annotations}
}
\description{
Performs gene set enrichment on numeric protein vector with specified Ontology
}
