% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust.R
\name{do_hclust}
\alias{do_hclust}
\title{Performs hierarchical clustering on data frame}
\usage{
do_hclust(
  data_,
  scale = T,
  group.column = "groups",
  grouping.function = mean,
  distance.method = "euclidean",
  clustering.method = "complete",
  input = "LFQ.intensity",
  output = "data_hclust"
)
}
\arguments{
\item{data_}{data list}

\item{scale}{Scale data (Z-score)?}

\item{group.column}{column to use for combining groups}

\item{grouping.function}{function to combine groups}

\item{distance.method}{method to calculate distance; see ?dist for options}

\item{clustering.method}{method to cluster columns and rows; see ?hclust for options}

\item{input}{name of data to use for analysis}
}
\value{

}
\description{
Performs hierarchical clustering on data frame
}
