% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust_enrich_T2G.R
\name{do_hclust_enrich_T2G}
\alias{do_hclust_enrich_T2G}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_hclust_enrich_T2G(
  data_,
  dend = "dend_y",
  max.k = 8,
  ...,
  dataset,
  input,
  output
)
}
\arguments{
\item{data_}{list or tibble}

\item{dend}{which dendogram to work with}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
