% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust_fun_enrich.R
\name{do_hclust_fun_enrich}
\alias{do_hclust_fun_enrich}
\title{Title}
\usage{
do_hclust_fun_enrich(
  data_,
  max.k = 8,
  simplify = F,
  database = "BP",
  input = "dend_y"
)
}
\arguments{
\item{data_}{}
}
\value{

}
\description{
Title
}
