% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust_transpose.R
\name{do_hclust_transpose}
\alias{do_hclust_transpose}
\title{Changes x and y dendrograms and transposes hclust data}
\usage{
do_hclust_transpose(
  data_,
  from.row.names,
  to.row.names,
  dataset,
  input = "data_hclust",
  output
)
}
\arguments{
\item{data_}{list or tibble}

\item{from.row.names}{row names column from initial data}

\item{to.row.names}{row names column after transposing}

\item{dataset}{dataset}

\item{input}{name of input data (must be "data_hclust)}

\item{output}{name of output data}
}
\value{

}
\description{
Changes x and y dendrograms and transposes hclust data
}
