% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ORA_GO.R
\name{do_ORA_GO}
\alias{do_ORA_GO}
\title{Performs over-representation analysis using Gene Ontology annotations}
\usage{
do_ORA_GO(
  proteins,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  database,
  dataset,
  view = F
)
}
\arguments{
\item{proteins}{numeric/logical vector of proteins indicating group}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction), "hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of proteins for annotation to be used for enrichment}

\item{maxGSSize}{maximum number of proteins for annotation to be used for enrichment}

\item{database}{which GO database to use}

\item{dataset}{dataset}

\item{view}{view results}
}
\value{

}
\description{
Performs over-representation analysis using Gene Ontology annotations
}
