% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ORA_KEGG.R
\name{do_ORA_KEGG}
\alias{do_ORA_KEGG}
\title{Performs over-representation analysis using KEGG annotations}
\usage{
do_ORA_KEGG(
  proteins,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  view = T,
  return.all = F,
  add.info = F
)
}
\arguments{
\item{proteins}{numeric/logical vector of proteins indicating group}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none","hochberg" (Benjamini-Hochberg correction), "bonferroni", "holm", "hommel", "BH", "BY", "fdr"}

\item{view}{view results}

\item{return.all}{return enrichResult object; useful for further analysis of enrichment results}

\item{add.info}{add additional information to the results data frame}
}
\value{

}
\description{
Performs over-representation analysis using KEGG annotations
}
