% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ORA_TERM2GENE.R
\name{do_ORA_TERM2GENE}
\alias{do_ORA_TERM2GENE}
\title{Performs over-representation analysis using Gene Ontology annotations}
\usage{
do_ORA_TERM2GENE(
  proteins,
  database,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 3,
  maxGSSize = 500
)
}
\arguments{
\item{proteins}{numeric/logical vector of proteins indicating group}

\item{database}{TERM2GENE data frame}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction),
"hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}
}
\description{
Performs over-representation analysis using Gene Ontology annotations
}
