% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ORA_topGO.R
\name{do_ORA_topGO}
\alias{do_ORA_topGO}
\title{Performs overexpression enrichment on logical protein vector with specified GO ontology}
\usage{
do_ORA_topGO(
  proteins,
  ontology = "CC",
  algorithm = "classic",
  threshold = 0.05,
  dataset,
  add.info = F
)
}
\arguments{
\item{proteins}{character vector}

\item{ontology}{GO ontology}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{threshold}{p-value/confidence threshold to exclude terms}

\item{dataset}{dataset}

\item{add.info}{Add additional information (takes longer)}
}
\description{
Performs overexpression enrichment on logical protein vector with specified GO ontology
}
