% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pca.R
\name{do_pca}
\alias{do_pca}
\title{Performs Principal Component Analysis using the prcomp function}
\usage{
do_pca(data_, scale = T, data.name = "raw_data", plot = T, print.summary = F)
}
\arguments{
\item{data_}{data list}

\item{scale}{Scale data (Z-scores)?}

\item{data.name}{name of data to use for analysis}

\item{plot}{plot results}

\item{print.summary}{should summary be printed}
}
\value{

}
\description{
Performs Principal Component Analysis using the prcomp function
}
