% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_reverse_dend.R
\name{do_reverse_dend}
\alias{do_reverse_dend}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_reverse_dend(data_, which = "y", dataset, input = "data_hclust", output)
}
\arguments{
\item{data_}{list or tibble}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}

\item{...}{specific arguments}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
