% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_something.R
\name{do_something}
\alias{do_something}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_something(data_, FUN, ..., dataset, input, output)
}
\arguments{
\item{data_}{list or tibble}

\item{FUN}{function to apply to data}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
