% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_expression.R
\name{do_t.test}
\alias{do_t.test}
\title{Calculates Fold change and p-value from given data frame or defined variables, observations and groups}
\usage{
do_t.test(
  data,
  groups,
  control,
  var.equal = T,
  paired = F,
  significance.method = "threshold",
  p.value.limit = 0.05,
  fold.change.abs.limit = 0,
  observations.set,
  dataset,
  plot = T,
  view = F,
  save = F
)
}
\arguments{
\item{data}{data frame}

\item{groups}{vector containing group information}

\item{control}{optional define control sample (case / control)}

\item{var.equal}{t-test with equal variance}

\item{paired}{t-test with paired samples}

\item{observations.set}{observations.set}

\item{dataset}{dataset}

\item{plot}{plot results}

\item{view}{view results matrix}

\item{save}{Save results matrix}
}
\value{

}
\description{
Calculates Fold change and p-value from given data frame or defined variables, observations and groups
}
