% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ttest.R
\name{do_ttest}
\alias{do_ttest}
\title{Performs t-test}
\usage{
do_ttest(
  data,
  group.column = "groups",
  control,
  case,
  paired = F,
  var.equal = T
)
}
\arguments{
\item{data}{data}

\item{group.column}{column to use as group identification}

\item{control}{control group}

\item{case}{case group}

\item{paired}{a logical indicating whether you want a paired t-test (see ?t.test)}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal (see ?t.test)}
}
\value{

}
\description{
Performs t-test
}
