% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_attributes.R
\name{.add_data_attributes}
\alias{.add_data_attributes}
\title{Adds entries to data_attributes}
\usage{
.add_data_attributes(data, which, new_attr)
}
\arguments{
\item{data}{data object with specific attributes}

\item{which}{which attribute type}

\item{new_attr}{new attributes}
}
\value{

}
\description{
Adds entries to data_attributes
}
