% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dend_.R
\name{enrich_dend_}
\alias{enrich_dend_}
\title{Calculates enrichment for each cluster level}
\usage{
enrich_dend_(cor_list, proteins, n = 10, name, cut = T, plot = T)
}
\arguments{
\item{cor_list}{cor_list object}

\item{proteins}{annotated proteins or protein score vector}

\item{n}{max number of clusters}

\item{name}{name}

\item{plot}{Plot results?}
}
\value{

}
\description{
Calculates enrichment for each cluster level
}
