% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dend_fisher.R
\name{enrich_dend_fisher}
\alias{enrich_dend_fisher}
\title{Performs Fisher's exact test on all cluster levels of given dendrogram}
\usage{
enrich_dend_fisher(dend.table, annotated.proteins, n = 10, inverse = F)
}
\arguments{
\item{dend.table}{dendrogram represented as data frame}

\item{annotated.proteins}{vector of annotated proteins}

\item{n}{maximum number of clusters}
}
\value{

}
\description{
Performs Fisher's exact test on all cluster levels of given dendrogram
}
