% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dend_ks.R
\name{enrich_dend_ks}
\alias{enrich_dend_ks}
\title{Performs KS test on all cluster levels of given dendrogram}
\usage{
enrich_dend_ks(dend.table, protein.scores, n = 10)
}
\arguments{
\item{dend.table}{drndrogram represented as dataframe}

\item{protein.scores}{vector of protein scores}

\item{n}{maximum number of clusters}
}
\value{

}
\description{
Performs KS test on all cluster levels of given dendrogram
}
