% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_fisher_GO.R
\name{enrichment_fisher_GO}
\alias{enrichment_fisher_GO}
\title{Performs overexpression enrichment on logical protein vector with specified GO ontology}
\usage{
enrichment_fisher_GO(
  proteins,
  ontology = "CC",
  algorithm = "classic",
  threshold = 0.05,
  add.info = F
)
}
\arguments{
\item{proteins}{character vector}

\item{ontology}{GO ontology}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{threshold}{p-value/confidence threshold to exclude terms}

\item{add.info}{Add additional information (takes longer)}
}
\value{

}
\description{
Performs overexpression enrichment on logical protein vector with specified GO ontology
}
