% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_ks_GO.R
\name{enrichment_ks_GO}
\alias{enrichment_ks_GO}
\title{Performs gene set enrichment on numeric protein vector with specified Ontology}
\usage{
enrichment_ks_GO(proteins, ontology, algorithm)
}
\arguments{
\item{proteins}{numeric score vector}

\item{ontology}{GO ontology}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}
}
\value{

}
\description{
Performs gene set enrichment on numeric protein vector with specified Ontology
}
