% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult_handling.R
\name{enrichResult2data.frame}
\alias{enrichResult2data.frame}
\title{Extracts result data frame Cluster enrichResult object or list of objects}
\usage{
enrichResult2data.frame(x)
}
\arguments{
\item{x}{enrichResult or list of objects}
}
\value{

}
\description{
Extracts result data frame Cluster enrichResult object or list of objects
}
