% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_2w_anova_i.R
\name{eval_2w_anova_i}
\alias{eval_2w_anova_i}
\title{Classifies protein changes from two-way ANOVA with interaction}
\usage{
eval_2w_anova_i(analysis_list, p.value.threshold = 0.05, fc.threshold = 0.5)
}
\arguments{
\item{p.value.threshold}{threshold for p-values}

\item{fc.threshold}{absolute threshold for log2 fold-change}

\item{data}{data frame output of two-way ANOVA with interactions}
}
\description{
Classifies protein changes from two-way ANOVA with interaction
}
