% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_anova_.R
\name{eval_anova_}
\alias{eval_anova_}
\title{Evaluates data from ANOVA do_anova() for significant proteins}
\usage{
eval_anova_(
  data_,
  anova.pvalueCutoff = 0.05,
  anova.pAdjustMethod = "BH",
  tukey.pvalueCutoff = 0.05,
  tukey.pAdjustMethod = "BH",
  fcThreshold = 0.5
)
}
\arguments{
\item{data_}{data list}

\item{anova.pvalueCutoff}{threshold for ANOVA p-values}

\item{anova.pAdjustMethod}{method to adjust ANOVA p-values}

\item{tukey.pvalueCutoff}{threshold for Tukey p-values}

\item{tukey.pAdjustMethod}{method to adjust Tukey p-values}

\item{fcThreshold}{absolute fold-change threshold}
}
\value{

}
\description{
Evaluates data from ANOVA do_anova() for significant proteins
}
