% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data.R
\name{eval_data}
\alias{eval_data}
\title{Evaluates data cell-wise}
\usage{
eval_data(data_, expr, FUN, input = "Peptides", output)
}
\arguments{
\item{data_}{data list}

\item{expr}{expression to be applied to each cell (must contain x as variable, e.g. x > 2)}

\item{FUN}{function to be applied to each cell}

\item{input}{if data_ is list: name of data to use}

\item{output}{if data_ is list: name of output data to save in list under}
}
\value{

}
\description{
Evaluates data cell-wise
}
