% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data_.R
\name{eval_data_}
\alias{eval_data_}
\title{Evaluates data cell-wise}
\usage{
eval_data_(data_, expr, FUN, data.name = "raw_data")
}
\arguments{
\item{data_}{data list}

\item{expr}{expression to be applied to each cell (must contain x as variable)}

\item{FUN}{function to be applied to each cell}

\item{data.name}{name of data to use}
}
\value{

}
\description{
Evaluates data cell-wise
}
