% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data_grouped_var.R
\name{eval_data_grouped_var}
\alias{eval_data_grouped_var}
\title{Evaluates data column-wise and saves variables data}
\usage{
eval_data_grouped_var(
  data,
  groups,
  expr,
  variables = "default",
  observations = "default",
  data.name,
  observations.set,
  dataset,
  name,
  view = F,
  save = F
)
}
\arguments{
\item{data}{data}

\item{groups}{group definition}

\item{expr}{function(x) how columns should be evaluated}

\item{variables}{variables}

\item{observations}{observations}

\item{data.name}{data name}

\item{observations.set}{set of observations}

\item{dataset}{dataset}

\item{name}{name to be saved under}

\item{view}{View results matrix?}

\item{save}{Save results matrix?}
}
\value{

}
\description{
Evaluates data column-wise and saves variables data
}
