% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data_var.R
\name{eval_data_var}
\alias{eval_data_var}
\title{Evaluates data columnwise and saves variables data}
\usage{
eval_data_var(
  variables,
  observations,
  expr,
  name,
  observations.set,
  data.name,
  type,
  dataset,
  set.default = F,
  save = T,
  return = F
)
}
\arguments{
\item{variables}{variables}

\item{observations}{observations}

\item{expr}{function(x) how columns should be evaluated}

\item{name}{name of new data}

\item{observations.set}{set of observations}

\item{data.name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{set.default}{should new data be set as default}

\item{save}{save}

\item{return}{return}
}
\value{

}
\description{
Evaluates data columnwise and saves variables data
}
