% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_dend_enrich.R
\name{eval_dend_enrich}
\alias{eval_dend_enrich}
\title{Evalueates dendrogram enrichment}
\usage{
eval_dend_enrich(dend.enrich, cut = T)
}
\arguments{
\item{dend.enrich}{output from enrich_dend_fisher/_ks}
}
\value{

}
\description{
Evalueates dendrogram enrichment
}
