% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_expression.R
\name{eval_expression}
\alias{eval_expression}
\title{Evaluates expression analysis data containing p.value and log2.fc}
\usage{
eval_expression(
  data,
  significance.method = "threshold",
  p.value.limit = 0.05,
  fold.change.abs.limit = 0
)
}
\arguments{
\item{data}{expression analysis data}

\item{significance.method}{method to evaluate p-values}

\item{p.value.limit}{cutoff or FDR for p-value}

\item{fold.change.limit}{absolute limit for the fold change}
}
\value{

}
\description{
Evaluates expression analysis data containing p.value and log2.fc
}
