% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_DT_t_test.R
\name{export_DT_t.test}
\alias{export_DT_t.test}
\title{Formats table for nice output (R Markdown)}
\usage{
export_DT_t.test(data, order.by = "p.adjust", descending = F)
}
\arguments{
\item{data}{data from t.test}

\item{order.by}{order rows by given column}

\item{descending}{arrange in descending order}
}
\value{

}
\description{
Formats table for nice output (R Markdown)
}
