% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_RMD_fun_enrich.R
\name{export_RMD_fun_enrich}
\alias{export_RMD_fun_enrich}
\title{Template for functions that accept either a data frame or a list}
\usage{
export_RMD_fun_enrich(
  data_,
  view = F,
  input = "fun_enrich",
  output = "fun_enrich_DT"
)
}
\arguments{
\item{data_}{list or tibble}

\item{view}{view table}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
