% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_RMD_t_test.R
\name{export_RMD_t.test}
\alias{export_RMD_t.test}
\title{Template for functions that accept either a data frame or a list}
\usage{
export_RMD_t.test(
  data_,
  order.by = "p.adjust",
  descending = F,
  dataset,
  view = F,
  input = "data_t.test",
  output = "data_DT_t.test"
)
}
\arguments{
\item{data_}{list or tibble}

\item{order.by}{order rows by given column}

\item{descending}{arrange in descending order}

\item{view}{view table}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
