% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_.R
\name{expr_}
\alias{expr_}
\title{Builds expression analysis list}
\usage{
expr_(data, groups, scale = T, combine = "mean", observations.set, dataset)
}
\arguments{
\item{data}{data}

\item{groups}{groups}

\item{scale}{scale data}

\item{combine}{combine groups}

\item{observations.set}{which observations set}

\item{dataset}{dataset}
}
\value{

}
\description{
Builds expression analysis list
}
