% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MSigDB_category.R
\name{extract_MSigDB_category}
\alias{extract_MSigDB_category}
\title{Extracts category specific TERM2GENE data frames from local MSigDB data}
\usage{
extract_MSigDB_category(category, dataset)
}
\arguments{
\item{category}{category (see MSig_categories)}

\item{dataset}{dataset}
}
\value{

}
\description{
Extracts category specific TERM2GENE data frames from local MSigDB data
}
