% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_pvalue.R
\name{fisher_pvalue}
\alias{fisher_pvalue}
\title{Calculates Fisher's exact test and returns p-value}
\usage{
fisher_pvalue(a, b)
}
\arguments{
\item{a}{logical vector}

\item{b}{logical vector}
}
\value{

}
\description{
Calculates Fisher's exact test and returns p-value
}
