% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test.R
\name{fisher_test}
\alias{fisher_test}
\title{Calculates p-value from Fisher's exact test}
\usage{
fisher_test(proteins, annotated.proteins)
}
\arguments{
\item{proteins}{logical vector of all proteins}

\item{annotated.proteins}{proteins in annotated group}
}
\value{

}
\description{
Calculates p-value from Fisher's exact test
}
