% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich.R
\name{fun_enrich}
\alias{fun_enrich}
\title{Performs appropriate enrichment on given sets of proteins or values}
\usage{
fun_enrich(
  proteins,
  background = NULL,
  inverse = F,
  databases = "GO",
  algorithm = "weight01",
  threshold = 0.05,
  add.info = F,
  view = T
)
}
\arguments{
\item{proteins}{list of proteins/scores/protein groups}

\item{background}{(optional) background genes}

\item{inverse}{scores: enriches for higher scores if TRUE}

\item{databases}{databases to use}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{threshold}{p-value/confidence threshold to exclude terms}

\item{add.info}{Add additional information (takes longer)}

\item{view}{View results?}
}
\value{

}
\description{
Performs appropriate enrichment on given sets of proteins or values
}
