% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich_.R
\name{fun_enrich_}
\alias{fun_enrich_}
\title{Title}
\usage{
fun_enrich_(cor_list, name, k, l, databases = "CC")
}
\arguments{
\item{cor_list}{}

\item{name}{}

\item{k}{}

\item{l}{}

\item{databases}{}
}
\value{

}
\description{
Title
}
