% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich_t_test.R
\name{fun_enrich_t.test}
\alias{fun_enrich_t.test}
\title{Functional enrichment of t.test results}
\usage{
fun_enrich_t.test(
  data_,
  pValueThreshold = 0.05,
  log2fcThreshold = 0,
  ...,
  view = F,
  dataset,
  input = "data_t.test",
  output = "fun_enrich_t.test"
)
}
\arguments{
\item{data_}{list or tibble}

\item{pValueThreshold}{threshold for t.test p-value}

\item{log2fcThreshold}{abs. threshold for log2 fold-change}

\item{view}{view results}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Functional enrichment of t.test results
}
