% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2p.R
\name{g2p}
\alias{g2p}
\title{Translates gene symbols into protein accession IDs}
\usage{
g2p(genes, dataset)
}
\arguments{
\item{genes}{gene symbols}

\item{dataset}{dataset}
}
\value{

}
\description{
Translates gene symbols into protein accession IDs
}
