% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2protein.R
\name{gene2protein}
\alias{gene2protein}
\title{Converts gene names to UniProt accession numbers (NO MAPPING OR DIFFERENT SPECIES SUPPORTED YET)}
\usage{
gene2protein(genes)
}
\arguments{
\item{genes}{vector containing genes}
}
\value{

}
\description{
Converts gene names to UniProt accession numbers (NO MAPPING OR DIFFERENT SPECIES SUPPORTED YET)
}
