% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_name.R
\name{get_data_name}
\alias{get_data_name}
\title{Check data type and return data.name}
\usage{
get_data_name(name, type, dataset)
}
\arguments{
\item{name}{(optional) name of data}

\item{type}{type of data}

\item{dataset}{datset}
}
\value{

}
\description{
Check data type and return data.name
}
