% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_names.R
\name{get_data_names}
\alias{get_data_names}
\title{Prints or returns data names}
\usage{
get_data_names(type, dataset, print = T, return = F)
}
\arguments{
\item{type}{data type}

\item{dataset}{dataset}

\item{print}{print}

\item{return}{return}
}
\value{

}
\description{
Prints or returns data names
}
