% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_type.R
\name{get_data_type}
\alias{get_data_type}
\title{Check and get data type}
\usage{
get_data_type(type, dataset)
}
\arguments{
\item{type}{data type}

\item{dataset}{dataset}
}
\value{

}
\description{
Check and get data type
}
