% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_types.R
\name{get_data_types}
\alias{get_data_types}
\title{Prints or returns datatypes}
\usage{
get_data_types(dataset, print = T, return = F)
}
\arguments{
\item{dataset}{dataset/s}

\item{print}{print}

\item{return}{return}
}
\value{

}
\description{
Prints or returns datatypes
}
