% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Checks and returns correct dataset identifier}
\usage{
get_dataset(dataset)
}
\arguments{
\item{dataset}{dataset name or number}
}
\value{

}
\description{
Checks and returns correct dataset identifier
}
