% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_data_name.R
\name{get_default_data_name}
\alias{get_default_data_name}
\title{Return default data name}
\usage{
get_default_data_name(type, dataset)
}
\arguments{
\item{type}{data type}

\item{dataset}{datset}
}
\value{

}
\description{
Return default data name
}
