% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_data_type.R
\name{get_default_data_type}
\alias{get_default_data_type}
\title{Returns default data type}
\usage{
get_default_data_type(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{

}
\description{
Returns default data type
}
