% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_proteins.R
\name{get_disease_proteins}
\alias{get_disease_proteins}
\title{Returns disease associated proteins}
\usage{
get_disease_proteins(terms)
}
\arguments{
\item{terms}{disease terms}
}
\value{

}
\description{
Returns disease associated proteins
}
