% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_proteins_TRY.R
\name{get_GO_proteins_TRY}
\alias{get_GO_proteins_TRY}
\title{Returns GO terms from given terms}
\usage{
get_GO_proteins_TRY(terms)
}
\arguments{
\item{terms}{GO ids}
}
\value{

}
\description{
Returns GO terms from given terms
}
