% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Returns groups vector}
\usage{
get_groups(groups, control, observations, observations.set, dataset)
}
\arguments{
\item{groups}{groups}

\item{control}{(optional) which group is control}

\item{observations}{observations}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}
}
\value{

}
\description{
Returns groups vector
}
