% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identifiers.R
\name{get_identifiers}
\alias{get_identifiers}
\title{Finds feasible identifier columns}
\usage{
get_identifiers(x, sep)
}
\arguments{
\item{x}{data frame}

\item{sep}{separator}
}
\value{

}
\description{
Finds feasible identifier columns
}
