% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_data.R
\name{get_obs_data}
\alias{get_obs_data}
\title{Return observations data}
\usage{
get_obs_data(observations, which, observations.set, dataset)
}
\arguments{
\item{observations}{(optional) vector of observations or expression}

\item{which}{which observations data to pull}

\item{observations.set}{observations data frame}

\item{dataset}{dataset}
}
\value{

}
\description{
Return observations data
}
