% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_data.R
\name{get_observations_data}
\alias{get_observations_data}
\title{Return observations data}
\usage{
get_observations_data(name, observations, observations.set, dataset)
}
\arguments{
\item{name}{name of data}

\item{observations}{(optional) vector of observations or expression}

\item{observations.set}{observations data frame}

\item{dataset}{dataset}
}
\value{

}
\description{
Return observations data
}
