% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_set.R
\name{get_observations_set}
\alias{get_observations_set}
\title{Checks and returns correct observations set identifier}
\usage{
get_observations_set(dataset, observations.set)
}
\arguments{
\item{dataset}{dataset name or number}

\item{observations.set}{set of observations}
}
\value{

}
\description{
Checks and returns correct observations set identifier
}
