% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_sets.R
\name{get_observations_sets}
\alias{get_observations_sets}
\title{Prints or returns observations sets}
\usage{
get_observations_sets(dataset, print = T)
}
\arguments{
\item{dataset}{dataset/s}

\item{print}{print}
}
\value{

}
\description{
Prints or returns observations sets
}
