% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_data.R
\name{get_var_data}
\alias{get_var_data}
\title{Return variables data}
\usage{
get_var_data(variables, which, dataset)
}
\arguments{
\item{variables}{(optional) vector of variables or expression}

\item{which}{which variables data to pull}

\item{dataset}{dataset}
}
\value{

}
\description{
Return variables data
}
