% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Assembles variables}
\usage{
get_variables(dataset, variables)
}
\arguments{
\item{dataset}{dataset}

\item{variables}{variables defined by dplyr::filter}
}
\value{

}
\description{
Assembles variables
}
