% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_extract_legend.R
\name{gg_extract_legend}
\alias{gg_extract_legend}
\title{Extract legend from ggplot object}
\usage{
gg_extract_legend(p)
}
\arguments{
\item{p}{ggplot object}
}
\description{
Extract legend from ggplot object
}
