% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_legend_position.R
\name{gg_legend.position}
\alias{gg_legend.position}
\title{Removes or repositions legend of ggplot}
\usage{
gg_legend.position(p, remove = T, legend.position = "right")
}
\arguments{
\item{p}{ggplot object}

\item{remove}{remove legend}

\item{legend.position}{positoin of legend if remove = FALSE}
}
\value{

}
\description{
Removes or repositions legend of ggplot
}
