% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{\%and\%}
\alias{\%and\%}
\title{Combines logical vectors using and}
\usage{
a \%and\% b
}
\arguments{
\item{a}{logical vector 1}

\item{b}{logical vector 2}
}
\value{

}
\description{
Combines logical vectors using and
}
