% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{\%or\%}
\alias{\%or\%}
\title{Combines logical vectors using or}
\usage{
a \%or\% b
}
\arguments{
\item{a}{logical vector 1}

\item{b}{logical vector 2}
}
\value{

}
\description{
Combines logical vectors using or
}
