% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2raw_dataset.R
\name{import2raw_dataset}
\alias{import2raw_dataset}
\title{Extracts variables and grouped data from imported data frame}
\usage{
import2raw_dataset(import, identifier, add = T)
}
\arguments{
\item{import}{imported data frame}

\item{identifier}{columns to use as identifiers}

\item{add}{add raw_dataset to .info list}
}
\value{

}
\description{
Extracts variables and grouped data from imported data frame
}
