% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_anything.R
\name{import_anything}
\alias{import_anything}
\title{Imports any file type using the file extension and returns list}
\usage{
import_anything(files, nolist = F)
}
\arguments{
\item{files}{file paths}

\item{nolist}{returns data frame directly if only one file is imported}
}
\value{

}
\description{
Imports any file type using the file extension and returns list
}
