% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_datasets.R
\name{import_datasets}
\alias{import_datasets}
\title{Import files and stores in a list; changes wd if files come from new wd}
\usage{
import_datasets(files)
}
\arguments{
\item{files}{files to be imported}
}
\value{

}
\description{
Import files and stores in a list; changes wd if files come from new wd
}
