% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta.R
\name{import_fasta}
\alias{import_fasta}
\title{Imports fasta file and saves proteome}
\usage{
import_fasta(file)
}
\arguments{
\item{file}{fasta file path}
}
\value{

}
\description{
Imports fasta file and saves proteome
}
