% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_norm_.R
\name{impute_norm_}
\alias{impute_norm_}
\title{Imputes values based on normal deviation}
\usage{
impute_norm_(
  analysis_list,
  shift = 1.8,
  width = 0.3,
  seed = 123,
  target.data = "raw_data"
)
}
\arguments{
\item{analysis_list}{analysis list}

\item{shift}{shift of normal distribution in standard deviations}

\item{width}{width of normal distribution in standard deviations}

\item{seed}{seed for reproducibility}

\item{target.data}{name of data to be modified}
}
\value{

}
\description{
Imputes values based on normal deviation
}
