% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_groups.R
\name{include_groups}
\alias{include_groups}
\title{Adds group vector to data frame}
\usage{
include_groups(data, groups, control, observations.set, dataset)
}
\arguments{
\item{data}{data frame}

\item{groups}{groups}

\item{control}{(optional) which group is control}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}
}
\value{

}
\description{
Adds group vector to data frame
}
