% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_groups.R
\name{include_groups_}
\alias{include_groups_}
\title{Adds group vector to data frame}
\usage{
include_groups_(
  data_,
  groups,
  control,
  observations.set,
  dataset,
  input = "LFQ.intensity",
  output
)
}
\arguments{
\item{data_}{analysis list}

\item{groups}{groups}

\item{control}{(optional) which group is control}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{input}{input data}

\item{output}{output data}
}
\value{

}
\description{
Adds group vector to data frame
}
