% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_observations_data.R
\name{include_observations_data_}
\alias{include_observations_data_}
\title{Adds group vector to data frame}
\usage{
include_observations_data_(
  data_,
  name,
  column.name,
  observations.set,
  dataset,
  input = "LFQ.intensity",
  output
)
}
\arguments{
\item{data_}{data frame}

\item{name}{observations data}

\item{column.name}{name of new column}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{input}{data frame to be modified}

\item{output}{dataframe to return in list}
}
\value{

}
\description{
Adds group vector to data frame
}
