% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_type.R
\name{is_data_type}
\alias{is_data_type}
\title{Check if data type exists}
\usage{
is_data_type(type, dataset)
}
\arguments{
\item{type}{data type}

\item{dataset}{dataset}
}
\value{

}
\description{
Check if data type exists
}
