% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_first.R
\name{keep_first}
\alias{keep_first}
\title{Removes multiple entries separated by a character}
\usage{
keep_first(x, sep)
}
\arguments{
\item{x}{vector}

\item{sep}{separator}
}
\value{

}
\description{
Removes multiple entries separated by a character
}
