% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test}
\alias{ks_test}
\title{Calculates p-value from Fisher's exact test}
\usage{
ks_test(protein.scores, annotated.proteins)
}
\arguments{
\item{protein.scores}{numeric vector of all proteins scores}

\item{annotated.proteins}{proteins in annotated group}
}
\value{

}
\description{
Calculates p-value from Fisher's exact test
}
