% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_abbr_table.R
\name{latex_abbr_table}
\alias{latex_abbr_table}
\title{Prints latex table code to console (made for abbreviations table of geegeedown)}
\usage{
latex_abbr_table(table, line.start = "\\t\\t", sort.abc = T)
}
\arguments{
\item{table}{position of table file}

\item{line.start}{character to start lines}

\item{sort.abc}{sort table alphabetically}
}
\description{
Prints latex table code to console (made for abbreviations table of geegeedown)
}
