% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_UniProt.R
\name{load_UniProt}
\alias{load_UniProt}
\title{Loads the species specific UniProt.ws to the .GlobalEnv}
\usage{
load_UniProt(taxId = get_taxId())
}
\arguments{
\item{taxId}{taxonomy Id}
}
\value{

}
\description{
Loads the species specific UniProt.ws to the .GlobalEnv
}
