% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_df2list.R
\name{logical_df2list}
\alias{logical_df2list}
\title{Turns identification dataframe into list of variable entries}
\usage{
logical_df2list(data)
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
Turns identification dataframe into list of variable entries
}
