% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variables.R
\name{match_variables}
\alias{match_variables}
\title{Return variables by matching variables.data}
\usage{
match_variables(variables.data, match.to, unique.matches = T, dataset)
}
\arguments{
\item{variables.data}{named vector or list of variables data to match}

\item{match.to}{column provided variables.data should be matched to}

\item{unique.matches}{remove redundant matches?}

\item{dataset}{dataset}
}
\value{

}
\description{
Return variables by matching variables.data
}
