% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2tibble.R
\name{matrix2tibble}
\alias{matrix2tibble}
\title{Transforms matrix to tibble and adds column for rownames}
\usage{
matrix2tibble(data, row.names)
}
\arguments{
\item{data}{matrix with row names}

\item{row.names}{name for rownames vector}
}
\value{

}
\description{
Transforms matrix to tibble and adds column for rownames
}
