% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_enrichment_results.R
\name{merge_enrichment_results}
\alias{merge_enrichment_results}
\title{Combines enrichment results data frames}
\usage{
merge_enrichment_results(
  enrichment_list,
  order.by = "p.adjust",
  descending = F,
  append.name = T
)
}
\arguments{
\item{enrichment_list}{list where each entry is a enrichment results data frame}

\item{order.by}{column name by which to order}

\item{descending}{order in descending order}

\item{append.name}{}
}
\value{

}
\description{
Combines enrichment results data frames
}
