% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_enrichment_results_recursive.R
\name{merge_enrichment_results_recursive}
\alias{merge_enrichment_results_recursive}
\title{merges a nested list of enrichment results}
\usage{
merge_enrichment_results_recursive(enrichment_list)
}
\arguments{
\item{enrichment_list}{nested list of enrichment dataframes}
}
\value{

}
\description{
merges a nested list of enrichment results
}
