% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult_handling.R
\name{merge_enrichResults}
\alias{merge_enrichResults}
\title{Merges a (nested) list of enrichResult objects}
\usage{
merge_enrichResults(
  enrichResult_list,
  order.by = "p.adjust",
  descending = F,
  append.name = T
)
}
\arguments{
\item{enrichResult_list}{nested list of enrichment data frames}
}
\description{
Merges a (nested) list of enrichResult objects
}
