% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_annotations_data_DisGeNet.R
\name{new_annotations_data_DisGeNet}
\alias{new_annotations_data_DisGeNet}
\title{Builds annotation list from DisGeNet}
\usage{
new_annotations_data_DisGeNet(nodeSize = 5, save = T)
}
\arguments{
\item{save}{Save?}
}
\value{

}
\description{
Builds annotation list from DisGeNet
}
