% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_annotations_data_UniProt.R
\name{new_annotations_data_UniProt}
\alias{new_annotations_data_UniProt}
\title{Builds annotation list from UniProt data}
\usage{
new_annotations_data_UniProt(annotation = "GO-ID", dataset, save = T)
}
\arguments{
\item{annotation}{annotation type}

\item{dataset}{dataset}

\item{save}{Save?}
}
\value{

}
\description{
Builds annotation list from UniProt data
}
