% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cache_list.R
\name{new_cache_list}
\alias{new_cache_list}
\title{Creates new cache list}
\usage{
new_cache_list(replace = T, return = F)
}
\arguments{
\item{replace}{Should cache list be replaced if already existing}

\item{return}{Should indication whether new cache list was created be returned}
}
\value{
whether a new cache list was created
}
\description{
Creates new cache list
}
