% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_databases_list.R
\name{new_databases_list}
\alias{new_databases_list}
\title{Creates databases list}
\usage{
new_databases_list(replace = F)
}
\arguments{
\item{replace}{Should databases list be replaced if already existing}
}
\value{

}
\description{
Creates databases list
}
