% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dataset.R
\name{new_dataset}
\alias{new_dataset}
\title{Sorts raw data into data list with attributes}
\usage{
new_dataset(
  x,
  name,
  data.origin,
  species,
  load.UniProt.ws = T,
  return = F,
  identifier,
  data.types,
  min.similarity = 8,
  min.groupsize = 6
)
}
\arguments{
\item{x}{raw data frame}

\item{name}{name of dataset}

\item{data.origin}{Identify origin software of data}

\item{species}{protein origin (T or F or UniProt taxosnomy ID)}

\item{load.UniProt.ws}{(optional) Should UniProt data be loaded}

\item{return}{Should dataset be returned?}

\item{identifier}{(optional) specific vector of identifier column/s}

\item{data.types}{(optional) types of data to extract; uses defaults if not specified}

\item{min.similarity}{minimum similarity oi column names}

\item{min.groupsize}{minimum number of samples}
}
\value{

}
\description{
Sorts raw data into data list with attributes
}
