% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_datasets_list.R
\name{new_datasets_list}
\alias{new_datasets_list}
\title{Creates datasets list}
\usage{
new_datasets_list(replace = F, return = T)
}
\arguments{
\item{replace}{Should datasets list be replaced if already existing}

\item{return}{Should indication whether new datasets info was created be returned}
}
\value{

}
\description{
Creates datasets list
}
