% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_GOdata.R
\name{new_GOdata}
\alias{new_GOdata}
\title{Builds GO object from all variables in the study}
\usage{
new_GOdata(
  proteins,
  ontology = "CC",
  statistic = "fisher",
  nodeSize = 10,
  dataset,
  save = T
)
}
\arguments{
\item{proteins}{(optional) prein vector}

\item{ontology}{ontology (CC, BP, MF)}

\item{statistic}{statistic}

\item{nodeSize}{minimum genes per annotation}

\item{dataset}{dataset}

\item{save}{Save?}
}
\value{

}
\description{
Builds GO object from all variables in the study
}
