% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalizes and returns data}
\usage{
normalize(data, method = "none", data2)
}
\arguments{
\item{data}{data}

\item{method}{method to use}

\item{data2}{(optional) second data set from which normalization factors should be generated}
}
\value{

}
\description{
Normalizes and returns data
}
