% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2g.R
\name{p2g}
\alias{p2g}
\title{Returns gene symbol from known proteins in dataset}
\usage{
p2g(proteins, dataset)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{
Returns gene symbol from known proteins in dataset
}
