% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2n.R
\name{p2n}
\alias{p2n}
\title{Returns names from known proteins in dataset}
\usage{
p2n(proteins, dataset)
}
\arguments{
\item{proteins}{}

\item{dataset}{}
}
\value{

}
\description{
Returns names from known proteins in dataset
}
