% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_composition.R
\name{plot_bar_composition}
\alias{plot_bar_composition}
\title{Plots data as bar plot with ggplot2}
\usage{
plot_bar_composition(
  data_,
  TERMS = "mitochondrion",
  color,
  xlab = "Fraction of Proteins",
  ylab = "",
  legend.name = "",
  legend.position = "right",
  aspect.ratio = 1,
  view = T,
  input = "data_annotation_composition",
  output = "plot_annotation_composition"
)
}
\arguments{
\item{data_}{data_}

\item{TERMS}{TERMS to plot}

\item{color}{names vector of colors}

\item{xlab}{labels of x-axis}

\item{ylab}{label of y-axis}

\item{legend.name}{name of lagend}

\item{legend.position}{position of legend ("none", "right", "bottom", "top", "left")}

\item{aspect.ratio}{aspect ratio of x- and y-axis}

\item{view}{view plot after generation}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Plots data as bar plot with ggplot2
}
