% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ComplexHeatmap.R
\name{plot_ComplexHeatmap}
\alias{plot_ComplexHeatmap}
\title{Template for functions that accept either a data frame or a list}
\usage{
plot_ComplexHeatmap(
  data_,
  ...,
  dataset,
  view = T,
  input = "data_hclust",
  output = "plot_ComplexHeatmap"
)
}
\arguments{
\item{data_}{list or tibble}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{view}{print plot}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
