% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor_matrix.R
\name{plot_cor_matrix}
\alias{plot_cor_matrix}
\title{Template for functions that accept either a data frame or a list}
\usage{
plot_cor_matrix(
  data_,
  dataset,
  view = T,
  export = F,
  file = "cor_matrix.pdf",
  height = 3,
  input = "data_cor_matrix",
  output = "plot_cor_matrix"
)
}
\arguments{
\item{data_}{list or tibble}

\item{dataset}{dataset}

\item{view}{print plot}

\item{export}{export plot}

\item{file}{path to exported file}

\item{height}{height of pdf (inches)}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Template for functions that accept either a data frame or a list
}
