% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cormat.R
\name{plot_cormat}
\alias{plot_cormat}
\title{Plots correlation matrix from cor_list object}
\usage{
plot_cormat(cor_list, data = "adjacency", print = T)
}
\arguments{
\item{cor_list}{cor_list object}

\item{data}{data type to be plotted (adjacency, similarity or correlation)}

\item{print}{print plot to device}
}
\value{

}
\description{
Plots correlation matrix from cor_list object
}
