% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cormat_.R
\name{plot_cormat_}
\alias{plot_cormat_}
\title{Plots correlation matrix from cor_list object}
\usage{
plot_cormat_(cor_list, data = "adjacency")
}
\arguments{
\item{cor_list}{cor_list object}

\item{data}{data type to be plotted (adjacency, similarity or correlation)}
}
\value{

}
\description{
Plots correlation matrix from cor_list object
}
