% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cormat_dxy.R
\name{plot_cormat_dxy}
\alias{plot_cormat_dxy}
\title{Plots cormat with dendrograms}
\usage{
plot_cormat_dxy(data, dend_x, dend_y)
}
\arguments{
\item{data}{data}

\item{dend_x}{x-axis dendrogram}

\item{dend_y}{y-axis dendrogram}

\item{scale}{should data be scaled}

\item{transform}{should data be transformed}
}
\value{

}
\description{
Plots cormat with dendrograms
}
