% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cormat_dy_box.R
\name{plot_cormat_dy_box}
\alias{plot_cormat_dy_box}
\title{Plots cormat with y-axis dendrogram}
\usage{
plot_cormat_dy_box(data, dend_y)
}
\arguments{
\item{data}{data}

\item{dend_y}{y-axis dendrogram}

\item{scale}{should data be scaled}

\item{transform}{should data be transformed}
}
\value{

}
\description{
Plots cormat with y-axis dendrogram
}
