% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cormat_enrichment_.R
\name{plot_cormat_enrichment_}
\alias{plot_cormat_enrichment_}
\title{Plots cormat with y-axis dendrogram}
\usage{
plot_cormat_enrichment_(cor_list, name, print = T)
}
\arguments{
\item{cor_list}{cor_list object}

\item{name}{name of enrichment analysis}

\item{print}{print plot to device}
}
\value{

}
\description{
Plots cormat with y-axis dendrogram
}
