% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cormat_enrichment_labels.R
\name{plot_cormat_enrichment_labels}
\alias{plot_cormat_enrichment_labels}
\title{Plots cormat with y-axis dendrogram and labels on the right y-axis}
\usage{
plot_cormat_enrichment_labels(cor_list, name, labels, print = T)
}
\arguments{
\item{cor_list}{cor_list object}

\item{name}{name of enrichment analysis}

\item{labels}{proteins to label on y-axis}

\item{print}{print plot to device}
}
\value{

}
\description{
Plots cormat with y-axis dendrogram and labels on the right y-axis
}
