% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dend_enrich_lines_.R
\name{plot_dend_enrich_lines_}
\alias{plot_dend_enrich_lines_}
\title{Plots -log10 p-value profile of dendrogram enrichment from cor_list object}
\usage{
plot_dend_enrich_lines_(cor_list, name)
}
\arguments{
\item{cor_list}{cor_list object}

\item{name}{name of enrichment}
}
\value{

}
\description{
Plots -log10 p-value profile of dendrogram enrichment from cor_list object
}
