% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrich.R
\name{plot_enrich}
\alias{plot_enrich}
\title{TPlots lines and cormat enrichment}
\usage{
plot_enrich(cor_list, name, labels)
}
\arguments{
\item{cor_list}{cor_list_object}

\item{name}{nem of enrichment}

\item{labels}{proteins to label yaxis}
}
\value{

}
\description{
TPlots lines and cormat enrichment
}
