% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_barplot.R
\name{plot_enrichment_barplot}
\alias{plot_enrichment_barplot}
\title{Plots barplot as used by clusterProfiler}
\usage{
plot_enrichment_barplot(
  data,
  x = "Count",
  color = "p.adjust",
  showCategory = 8,
  font.size = 10,
  title = "",
  label_format = 20,
  inverse = F,
  silent = F
)
}
\arguments{
\item{data}{enrichResult}

\item{x}{one of 'Count' and 'GeneRatio'}

\item{color}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{showCategory}{number of categories to show}

\item{font.size}{font size}

\item{title}{plot title}

\item{label_format}{a numeric value sets wrap length, alternatively a custom function to format axis labels. by default wraps names longer that 30 characters}

\item{silent}{suppress error message and only return NULL}
}
\value{

}
\description{
Plots barplot as used by clusterProfiler
}
