% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_euler.R
\name{plot_euler}
\alias{plot_euler}
\title{Plots Euler diagram from dataframe}
\usage{
plot_euler(data, view = T)
}
\arguments{
\item{data}{dataframe}

\item{view}{view plot}
}
\value{

}
\description{
Plots Euler diagram from dataframe
}
