% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_bar.R
\name{plot_gg_bar}
\alias{plot_gg_bar}
\title{Plots data as bar plot with ggplot2}
\usage{
plot_gg_bar(data, x = "observations", y = "count")
}
\arguments{
\item{data}{data to plot}

\item{x}{column name of x axis labels}

\item{y}{numeric value to be represented}
}
\value{

}
\description{
Plots data as bar plot with ggplot2
}
