% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_bar_anova_eval.R
\name{plot_gg_bar_anova_eval}
\alias{plot_gg_bar_anova_eval}
\title{Plots stacked bar plot to indicate significantly changed proteins from eval_anova output}
\usage{
plot_gg_bar_anova_eval(
  data,
  columns,
  labels,
  x.axis.title = "Group",
  y.axis.title = "Proteins",
  text.size = 6,
  text.color = "black",
  y.axis.breaks = 1,
  axis.line.size = 0.5,
  axis.color = "black",
  axis.ticks.size = 0.3,
  axis.title.size = 8,
  axis.text.size = 6,
  aspect.ratio = 0.8
)
}
\arguments{
\item{data}{data}

\item{columns}{columns to use for plot}

\item{labels}{x-axis labels for columns}

\item{x.axis.title}{title of x-axis}

\item{y.axis.title}{title of y-axis}

\item{text.size}{size of text in points (5-8)}

\item{text.color}{color of text}

\item{y.axis.breaks}{break size between ticks of y-axis}

\item{axis.line.size}{width of axes lines}

\item{axis.color}{color of axes lines}

\item{axis.ticks.size}{width of axis ticks}

\item{axis.title.size}{size of axis title}

\item{axis.text.size}{size of axis labels}

\item{aspect.ratio}{y/x ratio}
}
\value{

}
\description{
Plots stacked bar plot to indicate significantly changed proteins from eval_anova output
}
