% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_bar_composition.R
\name{plot_gg_bar_composition}
\alias{plot_gg_bar_composition}
\title{Plots data as bar plot with ggplot2}
\usage{
plot_gg_bar_composition(
  data_,
  TERMS = "mitochondrion",
  color,
  xlab = "Fraction of Proteins",
  ylab = "",
  legend.name = "",
  legend.position = "right",
  aspect.ratio = 1,
  view = T,
  input = "data_annotation_composition",
  output = "plot_annotation_composition"
)
}
\arguments{
\item{data}{data to plot}

\item{x}{column name of x axis labels}

\item{y}{numeric value to be represented}
}
\value{

}
\description{
Plots data as bar plot with ggplot2
}
