% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_count_id_quant.R
\name{plot_gg_count_id_quant}
\alias{plot_gg_count_id_quant}
\title{Barplot of identified/quantified proteins per sample}
\usage{
plot_gg_count_id_quant(
  data_,
  color.id = "#6BAED6",
  color.quant = "#2171B5",
  labels,
  order.by,
  xlab = NULL,
  ylab = "Proteins",
  limit.y.top,
  aspect.ratio = 0.66,
  dataset,
  input = "data_count_id_quant",
  output = "plot_count_id_quant",
  View = T
)
}
\arguments{
\item{data_}{data_ list}

\item{color.id}{color for identified proteins bar}

\item{color.quant}{color for quantified proteins bar}

\item{labels}{column name for labels}

\item{order.by}{order.by factor levels of specified column}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{limit.y.top}{upper limit of y-axis}

\item{aspect.ratio}{y/x ratio of plot axis}

\item{input}{name of id_quant data}

\item{output}{name of plot data}

\item{View}{print plot}
}
\value{

}
\description{
Barplot of identified/quantified proteins per sample
}
