% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_heatmap_protein_overlap.R
\name{plot_gg_heatmap_protein_overlap}
\alias{plot_gg_heatmap_protein_overlap}
\title{Compares log2 fold-changes between two comparisons from an anova_eval data list}
\usage{
plot_gg_heatmap_protein_overlap(
  data_,
  x,
  y,
  font_size = 8,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14,
  aspect.ratio = 1,
  point.size = c(0.5, 0.5, 1),
  color.not = "grey",
  color.up = "red",
  color.down = "blue",
  color.mixed = "violet",
  x.axis.limits,
  y.axis.limits,
  x.axis.breaks = 1,
  y.axis.breaks = 1,
  custom.theme = theme_hjv_framed,
  ...
)
}
\arguments{
\item{data_}{data list from anova_eval}

\item{x}{comparison on x-axis}

\item{y}{comparison on y-axis}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{point.size}{vector specifying size of unsignificant, significant in 1, and significant in both proteins}

\item{color.not}{color of not significant proteins}

\item{color.up}{color of upregulated proteins}

\item{color.down}{color of downregulated proteins}

\item{color.mixed}{color of both up- and downregulated proteins}

\item{x.axis.limits}{set limit of x-axis manually}

\item{y.axis.limits}{sets limits of y-axis manually}

\item{x.axis.breaks}{distance between breaks of x-axis ticks}

\item{y.axis.breaks}{distance between breaks of y-axis ticks}

\item{custom.theme}{theme to use for plot}

\item{...}{arguments for custom.theme}

\item{x.threshold}{absolute value for fold-change threshold lines}

\item{y.threshold}{absolute value for fold-change threshold lines}

\item{plot.center}{position of plot center}
}
\value{

}
\description{
Compares log2 fold-changes between two comparisons from an anova_eval data list
}
