% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_heatmap_terms.R
\name{plot_gg_heatmap_terms}
\alias{plot_gg_heatmap_terms}
\title{Plots heatmap with dendrograms}
\usage{
plot_gg_heatmap_terms(
  data_,
  TERMS = "mitochondrion",
  TERM2GENE,
  observations.order,
  label.proteins = T,
  label.observations = T,
  label.terms = T,
  indicate.clusters = T,
  k,
  cluster.line.width = 0.5,
  term.lines.height = 0.5,
  term.lines.width = 0.9,
  color.term.lines = "#A18276",
  color.cluster.annotation = "#463730",
  ratio.hm = 3,
  abs.height = 6,
  rel.width.hm.terms = 1,
  rel.width.dend.y = 0.2,
  rel.width.labels.y = 0.3,
  rel.height.dend.x = 0.05,
  rel.height.labels.x = 0.2,
  custom.theme.terms = theme_hjv_heatmap_only,
  terms.distance.method = "euclidean",
  terms.clustering.method = "complete",
  export = T,
  file = "heatmap.pdf"
)
}
\arguments{
\item{data_}{data}

\item{label.proteins}{should proteins be labeled on the y-axis?}

\item{rel.width.dend.y}{relative width of y-axis dendrogram}

\item{rel.width.labels.y}{width of labels (must be optimized)}

\item{rel.height.dend.x}{relative height of x-axis dendrogram}

\item{export}{export plot as pdf}

\item{file}{}

\item{height}{plot height in inch for export}

\item{ratio}{y/x ratio of heat map}
}
\value{

}
\description{
Plots heatmap with dendrograms
}
