% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_pca.R
\name{plot_gg_pca}
\alias{plot_gg_pca}
\title{Plots PCA using ggplot2}
\usage{
plot_gg_pca(data_, x = "PC1", y = "PC2", color = "groups")
}
\arguments{
\item{data_}{data list}

\item{x}{Principal component to plot on x-axis}

\item{y}{Principal component to plot on y-axis}

\item{color}{column to use for coloring data points}
}
\value{

}
\description{
Plots PCA using ggplot2
}
