% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_volcano.R
\name{plot_gg_volcano}
\alias{plot_gg_volcano}
\title{Plots a volcano plot using ggplot2}
\usage{
plot_gg_volcano(
  data,
  p.value.cutoff = 0.05,
  pos.log2fc.cutoff = 0,
  neg.log2fc.cutoff = 0,
  highlight.variables,
  highlight.color = "black",
  x.axis.title = "log2 fold-change",
  y.axis.title = "-log10(p-value)",
  text.size = 6,
  text.color = "black",
  point.size = 2,
  point.alpha = 0.8,
  highlight.point.size = 3,
  highlight.point.alpha = 0.8,
  x.axis.breaks = 1,
  y.axis.breaks = 1,
  axis.line.size = 0.5,
  axis.color = "black",
  axis.ticks.size = 0.3,
  axis.title.size = 8,
  axis.text.size = 6
)
}
\arguments{
\item{highlight.variables}{variables to highlight by point.size}

\item{text.size}{size of text in points (5-8)}

\item{text.color}{color of text}

\item{point.size}{point size (0.5-2)}

\item{point.alpha}{transparency (0-1)}

\item{highlight.point.size}{size of point of highlighted variables}

\item{axis.text.size}{}
}
\value{

}
\description{
Plots a volcano plot using ggplot2
}
