% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{PLots heatmap}
\usage{
plot_heatmap(
  variables,
  observations,
  groups,
  observations.set,
  data.name,
  type,
  dataset,
  scale = T,
  grouping,
  clustering.method = "complete",
  variables.labels,
  observations.labels,
  variables.order = T,
  observations.order = T,
  ratio = 3
)
}
\arguments{
\item{variables}{variables (see get variables)}

\item{observations}{observations (see get observations)}

\item{groups}{groups}

\item{observations.set}{set of observations}

\item{data.name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{scale}{should data be scaled}

\item{grouping}{should data be grouped}

\item{clustering.method}{clustering method}

\item{variables.labels}{labels of variables}

\item{observations.labels}{labels of observations}

\item{variables.order}{should variables be ordered by dendrogram}

\item{observations.order}{should observations be ordered by dendrogram}

\item{ratio}{y/x-ratio of tiles}
}
\value{

}
\description{
PLots heatmap
}
