% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_discrete.R
\name{plot_heatmap_discrete}
\alias{plot_heatmap_discrete}
\title{Plots heatmap from discrete values}
\usage{
plot_heatmap_discrete(
  data,
  transform = T,
  clustering.method = "complete",
  order.y = T,
  order.x = T,
  colors,
  labels,
  legend.title = NULL,
  legend.text.align = 0,
  ratio = 1,
  label.fun
)
}
\arguments{
\item{data}{data frame with first column variables}

\item{transform}{should data be transformed?}

\item{clustering.method}{argument for hclust}

\item{order.y}{either T/F if axis should be reordered by dendrogram, or vector containing all axis labels in order}

\item{order.x}{either T/F if axis should be reordered by dendrogram, or vector containing all axis labels in order}

\item{colors}{named vector of colors with names as corresponding values}

\item{labels}{named vector of labels with names as corresponding values}

\item{legend.title}{title of legend}

\item{legend.text.align}{0 = left, 1 = right}

\item{ratio}{x-y-ratio of heatmap tiles}

\item{label.fun}{function to generate labels with}
}
\value{

}
\description{
Plots heatmap from discrete values
}
