% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_dxy.R
\name{plot_heatmap_dxy}
\alias{plot_heatmap_dxy}
\title{Plots heatmap with dendrograms}
\usage{
plot_heatmap_dxy(data_, dend_x = T, dend_y = T, transpose = T)
}
\arguments{
\item{dend_x}{x-axis dendrogram}

\item{dend_y}{y-axis dendrogram}

\item{transpose}{should data be transformed}

\item{data}{data}
}
\value{

}
\description{
Plots heatmap with dendrograms
}
