% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Lines plot fo rpublication in ggplot2}
\usage{
plot_lines(
  data,
  group.column = "observations",
  label.column,
  x.y.ratio,
  main.color = "grey",
  highlight.color = "red",
  highlight.proteins = NULL,
  variables.label.fun = p2g,
  alpha = 0.8,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{data}

\item{group.column}{name of group column}

\item{label.column}{name of column containing labels for proteins}

\item{x.y.ratio}{ratio of x- and y-axis steps}

\item{main.color}{color of lines}

\item{highlight.color}{color of highlighted protein lines}

\item{highlight.proteins}{vector giving containing proteins to highlight}

\item{variables.label.fun}{function to transform accession IDs to labels}

\item{alpha}{transparency of lines}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{

}
\description{
Lines plot fo rpublication in ggplot2
}
