% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nothing.R
\name{plot_nothing}
\alias{plot_nothing}
\title{Template for functions that accept either a data frame or a list}
\usage{
plot_nothing(
  data_,
  ...,
  dataset,
  view = T,
  export = F,
  file = "nothing.pdf",
  height = 3,
  input = "data_nothing",
  output = "plot_nothing"
)
}
\arguments{
\item{data_}{list or tibble}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{view}{print plot}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
