% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plots PCA with ggplot2}
\usage{
plot_pca(
  data_,
  x = "PC1",
  y = "PC2",
  include.variance = T,
  fill.by = "observations",
  fill,
  shape.by,
  shape = 21,
  point.size = 1,
  point.transparency = 1,
  custom.theme = theme_hjv_framed_no_axes,
  aspect.ratio = 1,
  plot.center,
  axis.unit.ratio,
  expand.x.axis = c(0, 0),
  expand.y.axis = c(0, 0),
  x.axis.breaks = 1,
  y.axis.breaks = 1,
  legend.title.fill = "",
  legend.title.shape = "",
  legend.position = "right",
  legend.rows,
  view = T,
  input = "data_pca",
  output = "plot_pca",
  ...
)
}
\arguments{
\item{data_}{data list}

\item{x}{Principal component to plot on x-axis}

\item{y}{Principal component to plot on y-axis}

\item{include.variance}{add variance to axis titles}

\item{fill.by}{column to use for coloring data points}

\item{fill}{fill}

\item{shape.by}{column to use for shape of data points}

\item{shape}{shape}

\item{point.size}{point size (pt)}

\item{point.transparency}{transpacency (0-1)}

\item{custom.theme}{theme to use for plot}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{plot.center}{vector for center of plot}

\item{axis.unit.ratio}{ratio between x- and y-axis units}

\item{expand.x.axis}{expand x.axis (see scale_x_continuous)}

\item{expand.y.axis}{expand x.axis (see scale_y_continuous)}

\item{legend.title.fill}{title of fill legend}

\item{legend.title.shape}{title of shape legend}

\item{legend.position}{position of legend}

\item{view}{view plot}

\item{input}{name of input data}

\item{output}{name of output data}

\item{...}{}

\item{x.axis.break.size}{distance between x-axis breaks}

\item{y.axis.break.size}{distance between y-axis breaks}
}
\value{

}
\description{
Plots PCA with ggplot2
}
