% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{Visualizes identification data as UpSet plot}
\usage{
plot_upset(data, order, view = T)
}
\arguments{
\item{data}{data frame}

\item{order}{order of samples}

\item{view}{view plot}
}
\value{

}
\description{
Visualizes identification data as UpSet plot
}
