% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plots a venn diagram form given data}
\usage{
plot_venn(
  x,
  name,
  dataset,
  save = F,
  export.table = F,
  return.export.table = F
)
}
\arguments{
\item{x}{table or list}

\item{name}{name}

\item{dataset}{dataset}

\item{save}{save venn diagram groups}

\item{export.table}{export table}

\item{return.export.table}{return exported table}
}
\value{

}
\description{
Plots a venn diagram form given data
}
