% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano plot with plotly}
\usage{
plot_volcano(
  x,
  y,
  markers = "default",
  size = 4,
  opacity = opacity,
  axis.title.x = "log2 fold-change",
  axis.title.y = "-log10 p-value",
  add.protein.function = F,
  print = T
)
}
\arguments{
\item{x}{log2 fold change}

\item{y}{p-value}

\item{markers}{specific proteins to highlight}

\item{size}{size of dots}

\item{opacity}{opacity of dots}

\item{axis.title.x}{x-axis title}

\item{axis.title.y}{y-axis title}

\item{add.protein.function}{add protein function as hover information}

\item{print}{print plot to device}
}
\value{

}
\description{
Plot volcano plot with plotly
}
