% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy.R
\name{plot_xy}
\alias{plot_xy}
\title{Matches names and plots scatter plot of two variables}
\usage{
plot_xy(x, y, axis.title.x = "x", axis.title.y = "y")
}
\arguments{
\item{x}{names x vector}

\item{y}{named y vector}

\item{axis.title.x}{title of x-axis}

\item{axis.title.y}{title of y-axis}
}
\value{

}
\description{
Matches names and plots scatter plot of two variables
}
