% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{prepare_datasets}
\alias{prepare_datasets}
\title{Prepares imported data}
\usage{
prepare_datasets(
  imports,
  find.taxonomy = T,
  identify.data.origin = T,
  load.UniProt.ws = T
)
}
\arguments{
\item{imports}{imported raw data}

\item{find.taxonomy}{should taxonomy be determined}

\item{identify.data.origin}{Identify origin software of data}

\item{load.UniProt.ws}{Should a UniProt database be downloaded}
}
\value{

}
\description{
Prepares imported data
}
