% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein2taxId.R
\name{protein2taxID}
\alias{protein2taxID}
\title{Checks protein Ids for their species}
\usage{
protein2taxID(x, try = 1, silent = F)
}
\arguments{
\item{x}{vector of identifiers}

\item{try}{counter}

\item{silent}{should message be printed when Id found}
}
\value{

}
\description{
Checks protein Ids for their species
}
