% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_variables_data.R
\name{pull_variables_data}
\alias{pull_variables_data}
\title{Return variables data}
\usage{
pull_variables_data(name, variables, dataset, vector.names = "variables")
}
\arguments{
\item{name}{name of data}

\item{variables}{(optional) vector of variables or expression}

\item{dataset}{dataset}

\item{vector.names}{column to use as names for vector}
}
\value{

}
\description{
Return variables data
}
