% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_count_id_quant.R
\name{quick_protein_count}
\alias{quick_protein_count}
\title{Wrapper around do_count_id_quant}
\usage{
quick_protein_count(variables, observations, observations.set, dataset)
}
\arguments{
\item{variables}{variables}

\item{observations}{observations}

\item{observations.set}{observations set}

\item{dataset}{dataset}
}
\value{

}
\description{
Wrapper around do_count_id_quant
}
