% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_variables_threshold.R
\name{remove_variables_threshold}
\alias{remove_variables_threshold}
\title{Removes columns from data frame based on minimum values above 0}
\usage{
remove_variables_threshold(data, min = 0.5)
}
\arguments{
\item{data}{data}

\item{min}{min. fraction of values above 0}
}
\value{

}
\description{
Removes columns from data frame based on minimum values above 0
}
