% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_variables_threshold_.R
\name{remove_variables_threshold_}
\alias{remove_variables_threshold_}
\title{Removes columns from data frame based on minimum values above 0}
\usage{
remove_variables_threshold_(analysis_list, min = 0.5, target.data = "raw_data")
}
\arguments{
\item{analysis_list}{analysis list}

\item{min}{min. fraction of values above 0}

\item{target.data}{data frame to be modified}
}
\value{

}
\description{
Removes columns from data frame based on minimum values above 0
}
