% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_.R
\name{scale_}
\alias{scale_}
\title{Scales data form tibbles or analysis_list containing a tibble}
\usage{
scale_(analysis_list, data.name = "data", new.name = "data")
}
\arguments{
\item{analysis_list}{data list}

\item{data.name}{name of data entry}

\item{new.name}{new name of scaled data}
}
\value{

}
\description{
Scales data form tibbles or analysis_list containing a tibble
}
