% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2output.R
\name{select2output}
\alias{select2output}
\title{Formats mapping from select methods}
\usage{
select2output(output.data, keys, output, database, dataset, ...)
}
\arguments{
\item{output.data}{mapping data from select method}

\item{keys}{keys}

\item{output}{output type}

\item{database}{database of origin ("org" or "UniProt")}

\item{dataset}{dataset}

\item{...}{arguments for TERM2GENE function (subtype = "CC", "BP", or "MF")}
}
\description{
Formats mapping from select methods
}
