% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_DisGeNet.R
\name{select_DisGeNet}
\alias{select_DisGeNet}
\title{Improved select function for DisGeNet databases}
\usage{
select_DisGeNet(genes)
}
\arguments{
\item{genes}{vector of genes}
}
\description{
Improved select function for DisGeNet databases
}
