% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fasta.R
\name{select_fasta}
\alias{select_fasta}
\title{Retrieve annotation data from imported fasta files}
\usage{
select_fasta(keys, columns, output = "vector.keep", keytype, dataset)
}
\arguments{
\item{keys}{vector of protein identifiers (see keytypes_org())}

\item{columns}{data columns to return(see columns_org())}

\item{output}{format of output ("vector.keep" (default), "vector.na, "vector.rm", "mapping.na", "mapping.rm", "TERM2GENE")}

\item{keytype}{Id type of supplied keys (identified automatically if not provided)}

\item{dataset}{dataset}
}
\value{

}
\description{
Retrieve annotation data from imported fasta files
}
