% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_org_2_TERM2GENE.R
\name{select_org_2_TERM2GENE}
\alias{select_org_2_TERM2GENE}
\title{Exports TERM2GENE dataframe from select_org}
\usage{
select_org_2_TERM2GENE(mapping, subtype, dataset)
}
\arguments{
\item{mapping}{output from select_org}

\item{subtype}{subtype of mapping; e.g. Ontology}

\item{dataset}{dataset}
}
\description{
Exports TERM2GENE dataframe from select_org
}
