% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_UniProt.R
\name{select_UniProt}
\alias{select_UniProt}
\title{Improved select function for UniProt databases}
\usage{
select_UniProt(keys, columns, keytype = "UNIPROTKB", x, modify = T, dataset)
}
\arguments{
\item{keys}{keys}

\item{columns}{data to query}

\item{keytype}{type of key}

\item{x}{UniProt data base}

\item{modify}{(recommended) Should output values be modified to to singular
entries?}

\item{dataset}{dataset}
}
\value{

}
\description{
Improved select function for UniProt databases
}
