% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_UniProt_2_TERM2GENE.R
\name{select_UniProt_2_TERM2GENE}
\alias{select_UniProt_2_TERM2GENE}
\title{Exports TERM2GENE dataframe from select_org}
\usage{
select_UniProt_2_TERM2GENE(mapping)
}
\arguments{
\item{mapping}{output from select_UniProt}
}
\value{

}
\description{
Exports TERM2GENE dataframe from select_org
}
