% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_data_name.R
\name{set_default_data_name}
\alias{set_default_data_name}
\title{Set default data name}
\usage{
set_default_data_name(name, type, dataset, silent = T)
}
\arguments{
\item{name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{silent}{Silent?}
}
\value{

}
\description{
Set default data name
}
