% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_data_type.R
\name{set_default_data_type}
\alias{set_default_data_type}
\title{Set default data type}
\usage{
set_default_data_type(type, dataset, silent = T)
}
\arguments{
\item{type}{data type}

\item{dataset}{dataset}

\item{silent}{Silent?}
}
\value{

}
\description{
Set default data type
}
