% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wd.R
\name{set_wd}
\alias{set_wd}
\title{Changes working directory}
\usage{
set_wd(wd, save = T)
}
\arguments{
\item{wd}{working directory}

\item{save}{Should new working directory be saved?}
}
\value{

}
\description{
Changes working directory
}
