% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_annotations.R
\name{setup_annotations}
\alias{setup_annotations}
\title{Sets up annotaion database for biological Id translation and functional enrichment}
\usage{
setup_annotations(dataset, taxId, OrgDb)
}
\arguments{
\item{dataset}{dataset}

\item{taxId}{taxonomy Id}

\item{OrgDb}{name of annotation package from Bioconductor}
}
\value{

}
\description{
Sets up annotaion database for biological Id translation and functional enrichment
}
