% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_CORUM_annotations.R
\name{setup_CORUM_annotations}
\alias{setup_CORUM_annotations}
\title{Downloads CORUM database for functional enrichment}
\usage{
setup_CORUM_annotations()
}
\description{
Downloads CORUM database for functional enrichment
}
