% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_DisGeNet.R
\name{setup_DisGenet}
\alias{setup_DisGenet}
\title{Registers API key to use disgenet2r (https://www.disgenet.org/signup/)}
\usage{
setup_DisGenet(email = "nhutt069@uottawa.ca", password = "makey0ur0wnacc0unt")
}
\arguments{
\item{email}{email for DisGeNet account}

\item{password}{password for DisGeNet account}
}
\value{

}
\description{
Registers API key to use disgenet2r (https://www.disgenet.org/signup/)
}
