% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_EV_TOP_100.R
\name{setup_EV_TOP_100}
\alias{setup_EV_TOP_100}
\title{Downloads Top 100 EV protein markers from Vesiclepedia}
\usage{
setup_EV_TOP_100(id = "EV_TOP_100", type = "EV_database")
}
\value{

}
\description{
Downloads Top 100 EV protein markers from Vesiclepedia
}
