% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setup_folders}
\alias{setup_folders}
\title{Creates folder strucure for analysis project}
\usage{
setup_folders(dir, silent = F)
}
\arguments{
\item{dir}{directory of project}

\item{silent}{suppress message when folders created}
}
\value{

}
\description{
Creates folder strucure for analysis project
}
