% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fun_enrich.R
\name{setup_fun_enrich}
\alias{setup_fun_enrich}
\title{Initiates databases for functional enrichment}
\usage{
setup_fun_enrich(databases = "all")
}
\arguments{
\item{databases}{databases to setup}
}
\value{

}
\description{
Initiates databases for functional enrichment
}
