% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_MSigDB.R
\name{setup_MSigDB}
\alias{setup_MSigDB}
\title{Downloads and prepares MSigDB annotations}
\usage{
setup_MSigDB(dataset, category)
}
\arguments{
\item{dataset}{dataset}

\item{category}{categories to extract for enrichment}
}
\value{

}
\description{
Downloads and prepares MSigDB annotations
}
