% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_org_database.R
\name{setup_org_database}
\alias{setup_org_database}
\title{Sets up annotaion database for biological Id translation and functional enrichment}
\usage{
setup_org_database(dataset, OrgDb, taxId)
}
\arguments{
\item{dataset}{dataset}

\item{OrgDb}{name of annotation package from Bioconductor}

\item{taxId}{taxonomy Id}
}
\description{
Sets up annotaion database for biological Id translation and functional enrichment
}
