% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_STRINGdb.R
\name{setup_STRINGdb}
\alias{setup_STRINGdb}
\title{Sets up the STRING database}
\usage{
setup_STRINGdb(version = "11", score_threshold = 200, dataset)
}
\arguments{
\item{version}{STRING database version}

\item{dataset}{}
}
\value{

}
\description{
Sets up the STRING database
}
