% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_UniProt.R
\name{setup_UniProt}
\alias{setup_UniProt}
\title{Loads the species specific UniProt.ws to the .GlobalEnv}
\usage{
setup_UniProt(taxId, dataset)
}
\arguments{
\item{taxId}{taxonomy Id}

\item{dataset}{dataset}
}
\description{
Loads the species specific UniProt.ws to the .GlobalEnv
}
