% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_observations.R
\name{show_observations}
\alias{show_observations}
\title{Prints observations data frame}
\usage{
show_observations(observations.set, dataset)
}
\arguments{
\item{observations.set}{observations set}

\item{dataset}{dataset}
}
\value{

}
\description{
Prints observations data frame
}
