% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_variables.R
\name{show_variables}
\alias{show_variables}
\title{Prints variables data frame}
\usage{
show_variables(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{

}
\description{
Prints variables data frame
}
