% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{calculates similarity from correlation}
\usage{
similarity(data, method)
}
\arguments{
\item{data}{correlation matrix}

\item{method}{similarity function}
}
\value{

}
\description{
calculates similarity from correlation
}
