% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_.R
\name{similarity_}
\alias{similarity_}
\title{Adds similarity to cor_list}
\usage{
similarity_(cor_list, method = "none")
}
\arguments{
\item{cor_list}{cor_list object}

\item{method}{similarity method}
}
\value{

}
\description{
Adds similarity to cor_list
}
