% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_.R
\name{strsplit_}
\alias{strsplit_}
\title{Same as strsplit but returns vector}
\usage{
strsplit_(string, split = ";")
}
\arguments{
\item{string}{string}

\item{split}{split character}
}
\value{

}
\description{
Same as strsplit but returns vector
}
