% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TERM2GENE_2_list.R
\name{TERM2GENE_2_list}
\alias{TERM2GENE_2_list}
\title{Splits TERM2GENE dataframe to list}
\usage{
TERM2GENE_2_list(TERM2GENE)
}
\arguments{
\item{TERM2GENE}{TERM2GENE dataframe}
}
\value{

}
\description{
Splits TERM2GENE dataframe to list
}
