% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TERMS_in_proteins.R
\name{TERMS_in_proteins}
\alias{TERMS_in_proteins}
\title{Returns logical dataframe which proteins is annotated by given terms}
\usage{
TERMS_in_proteins(proteins, TERMS, TERM2GENE)
}
\arguments{
\item{proteins}{}

\item{TERMS}{}

\item{TERM2GENE}{}
}
\description{
Returns logical dataframe which proteins is annotated by given terms
}
