% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_small.R
\name{theme_small}
\alias{theme_small}
\title{Theme for small publication figures}
\usage{
theme_small(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{CHANGE}
}
\value{

}
\description{
Theme for small publication figures
}
