% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_0.R
\name{threshold_0}
\alias{threshold_0}
\title{Removes columns from data frame based on minimum values above 0}
\usage{
threshold_0(data, min = 0.5, groups)
}
\arguments{
\item{data}{data}

\item{min}{min. fraction of values above 0}

\item{groups}{(optional) groups}
}
\value{

}
\description{
Removes columns from data frame based on minimum values above 0
}
