% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_helpers.R, R/tibble2data.frame.R
\name{tibble2data.frame}
\alias{tibble2data.frame}
\title{Transforms tibble to data frame}
\usage{
tibble2data.frame(tibble, to.row.names)

tibble2data.frame(tibble, to.row.names)
}
\arguments{
\item{tibble}{tibble}

\item{to.row.names}{column to use as new row names}
}
\value{



}
\description{
Transforms tibble to data frame

Transforms tibble to data frame
}
