% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble2matrix.R
\name{tibble2matrix}
\alias{tibble2matrix}
\title{Transforms tibble to matrix}
\usage{
tibble2matrix(tibble, row.names)
}
\arguments{
\item{tibble}{tibble}

\item{row.names}{row to use for row names}
}
\value{

}
\description{
Transforms tibble to matrix
}
