% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_proteins_summary.R
\name{total_proteins_summary}
\alias{total_proteins_summary}
\title{INCOMOPLETE FUNCTION: Combines data frames with variables info}
\usage{
total_proteins_summary(input, dataset)
}
\arguments{
\item{input}{list of protein data frames}

\item{dataset}{dataset}
}
\value{

}
\description{
INCOMOPLETE FUNCTION: Combines data frames with variables info
}
