% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data_frames.R
\name{transfer_data_frames}
\alias{transfer_data_frames}
\title{Transfers data from raw_dataset to .datasets}
\usage{
transfer_data_frames(name, data.type)
}
\arguments{
\item{name}{dataset name}

\item{data.type}{data frames to transfer}
}
\value{

}
\description{
Transfers data from raw_dataset to .datasets
}
