% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_variables_data.R
\name{transfer_variables_data}
\alias{transfer_variables_data}
\title{Transfers variables data from raw_dataset to .datasets}
\usage{
transfer_variables_data(name, data.columns)
}
\arguments{
\item{name}{dataset name}

\item{data.columns}{columns to transfer}
}
\value{

}
\description{
Transfers variables data from raw_dataset to .datasets
}
