% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_data.R
\name{transpose_data}
\alias{transpose_data}
\title{Transposes tibble and uses first column as column names}
\usage{
transpose_data(data_, from.row.names, to.row.names, input, output)
}
\arguments{
\item{data_}{data_ list}

\item{from.row.names}{row names column of initial data frame}

\item{to.row.names}{row names column after transposing}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Transposes tibble and uses first column as column names
}
