% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_data_.R
\name{transpose_data_}
\alias{transpose_data_}
\title{Transposes tibble and uses first column as column names}
\usage{
transpose_data_(
  data_,
  from.row.names = "observations",
  to.row.names = "variables",
  input = "Peptides",
  output
)
}
\arguments{
\item{data_}{data_ list}

\item{from.row.names}{row names column of initial data frame}

\item{to.row.names}{row names column after transposing}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Transposes tibble and uses first column as column names
}
