% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_names.R
\name{update_data_names}
\alias{update_data_names}
\title{Add new data name}
\usage{
update_data_names(name, type, dataset, set.default = T)
}
\arguments{
\item{name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{set.default}{Should data name be set as default}
}
\value{

}
\description{
Add new data name
}
