% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_types.R
\name{update_data_types}
\alias{update_data_types}
\title{Adds new data type}
\usage{
update_data_types(dataset, default.type = NA)
}
\arguments{
\item{dataset}{dataset}

\item{default.type}{new default data type}
}
\value{

}
\description{
Adds new data type
}
