% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_datasets.R
\name{update_datasets}
\alias{update_datasets}
\title{Update list of datasets}
\usage{
update_datasets(name, set.default = F)
}
\arguments{
\item{name}{name of new dataset}
}
\value{

}
\description{
Update list of datasets
}
