% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_default_dataset.R
\name{update_default_dataset}
\alias{update_default_dataset}
\title{Updates default dataset}
\usage{
update_default_dataset(name, silent = F)
}
\arguments{
\item{name}{Name of dataset}

\item{silent}{Silent?}
}
\value{

}
\description{
Updates default dataset
}
