% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_observations_sets.R
\name{update_observations_sets}
\alias{update_observations_sets}
\title{Adds observations set}
\usage{
update_observations_sets(set, dataset, set.default = T)
}
\arguments{
\item{set}{observations set}

\item{dataset}{dataset}

\item{set.default}{should default be changed}
}
\value{

}
\description{
Adds observations set
}
