% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_default_attr.R
\name{view_default_attr}
\alias{view_default_attr}
\title{Prints default dataset attributes}
\usage{
view_default_attr(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{

}
\description{
Prints default dataset attributes
}
