% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_observations.R
\name{view_observations}
\alias{view_observations}
\title{Prints observations data frame}
\usage{
view_observations(observations.set, dataset)
}
\arguments{
\item{observations.set}{observations set}

\item{dataset}{dataset}
}
\value{

}
\description{
Prints observations data frame
}
