% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_proteins.R
\name{view_proteins}
\alias{view_proteins}
\title{Views a list of proteins}
\usage{
view_proteins(proteins)
}
\arguments{
\item{proteins}{vector of proteins}
}
\value{

}
\description{
Views a list of proteins
}
