% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_variables.R
\name{view_variables}
\alias{view_variables}
\title{Prints variables data frame}
\usage{
view_variables(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{

}
\description{
Prints variables data frame
}
