% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Adds new data}
\usage{
add_data(data, name, type, dataset, set.default = F)
}
\arguments{
\item{data}{new data}

\item{name}{data name}

\item{type}{data type}

\item{dataset}{dataset}

\item{set.default}{Set new data as default}
}
\value{

}
\description{
Adds new data
}
