% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data2dataset.R
\name{add_data2dataset}
\alias{add_data2dataset}
\title{Add variables data and data in matrix form to a dataset}
\usage{
add_data2dataset(x, dataset, data.types, data.origin)
}
\arguments{
\item{x}{raw data}

\item{dataset}{dataset list}

\item{data.types}{types of data to be imported}

\item{data.origin}{origin of data}
}
\value{

}
\description{
Add variables data and data in matrix form to a dataset
}
