% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_t.test.R
\name{add_observations_data}
\alias{add_observations_data}
\title{Adds data to observations dataframe}
\usage{
add_observations_data(
  data,
  name,
  observations.set,
  dataset,
  ignore.names = F,
  add.background.variable = T,
  replace
)
}
\arguments{
\item{data}{new data}

\item{name}{name}

\item{observations.set}{set of observations}

\item{dataset}{dataset}

\item{ignore.names}{Assumes that data matches observations}

\item{add.background.variable}{add name of column to the background variables for easy access}

\item{replace}{replace existing column}
}
\value{

}
\description{
Adds data to observations dataframe
}
