% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_observations_data.R
\name{add_observations_data}
\alias{add_observations_data}
\title{Adds data to observations dataframe}
\usage{
add_observations_data(data, name, dataset, observations.set)
}
\arguments{
\item{data}{new data}

\item{name}{name}

\item{dataset}{dataset}

\item{observations.set}{set of observations}
}
\value{

}
\description{
Adds data to observations dataframe
}
