% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables_data.R
\name{add_variables_data}
\alias{add_variables_data}
\title{Add data to variables}
\usage{
add_variables_data(data, name, dataset, set.default = F, replace)
}
\arguments{
\item{data}{new data}

\item{name}{name}

\item{dataset}{dataset}

\item{set.default}{set new variables data as default}

\item{replace}{replace existing column}
}
\value{

}
\description{
Add data to variables
}
