% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_protein_identification.R
\name{classify_protein_identification_2}
\alias{classify_protein_identification_2}
\title{Classifies proteins based on identification threshold}
\usage{
classify_protein_identification_2(
  data.eval,
  upper.neg = 0,
  lower.pos = 0.5,
  plot = T,
  export = F,
  save = F,
  return = T
)
}
\arguments{
\item{data.eval}{data from eval_ function}

\item{upper.neg}{upper limit for "not identified"}

\item{lower.pos}{lower limit for "confidently identified"}

\item{plot}{Plot venn diagram?}

\item{export}{Export ven diagram plot?}

\item{save}{Save results list?}

\item{return}{Return results list?}
}
\value{

}
\description{
Classifies proteins based on identification threshold
}
