% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_groups.R
\name{collapse_groups}
\alias{collapse_groups}
\title{Collapses groups of data frame by specified function and adds as list entry}
\usage{
collapse_groups(
  data_,
  FUN = mean,
  group.column = "groups",
  input = "Peptides",
  output
)
}
\arguments{
\item{data_}{data}

\item{FUN}{function to combine groups}

\item{group.column}{column specifying the groups}

\item{input}{name of input data frame}

\item{new.name}{name of output data frame (defaults to input data name)}
}
\value{

}
\description{
Collapses groups of data frame by specified function and adds as list entry
}
