% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_.R
\name{connectivity_}
\alias{connectivity_}
\title{Add connectivity and intracluster connectivity to cor_list}
\usage{
connectivity_(cor_list, k = 20, scale = T)
}
\arguments{
\item{cor_list}{cor_list object}

\item{k}{max number of clusters}

\item{scale}{Scale connectivity values}
}
\value{

}
\description{
Add connectivity and intracluster connectivity to cor_list
}
