% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram2table_.R
\name{dendrogram2table_}
\alias{dendrogram2table_}
\title{Adds table representation of dendrogram}
\usage{
dendrogram2table_(cor_list)
}
\arguments{
\item{cor_list}{cor_list object}
}
\value{

}
\description{
Adds table representation of dendrogram
}
