% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_1w_anova.R
\name{do_1w_anova}
\alias{do_1w_anova}
\title{Performs one-way ANOVA}
\usage{
do_1w_anova(
  analysis_list,
  formula = "log2(x) ~ groups",
  data.name = "raw_data"
)
}
\arguments{
\item{analysis_list}{list containing raw_data}

\item{formula}{formula given as string}

\item{data.name}{data to use for analysis}
}
\value{

}
\description{
Performs one-way ANOVA
}
