% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_count_id_quant.R
\name{do_count_id_quant}
\alias{do_count_id_quant}
\title{Combines information of identified and quantified proteins}
\usage{
do_count_id_quant(
  data_,
  input.id = "Unique.peptides",
  input.quant = "LFQ.intensity",
  output = "data_count_id_quant"
)
}
\arguments{
\item{data_}{data_}

\item{input.id}{data frame of identified proteins}

\item{input.quant}{data frame of quantified proteins}

\item{output}{name of output data}
}
\description{
Combines information of identified and quantified proteins
}
