% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GSEA_KEGG.R
\name{do_GSEA_KEGG}
\alias{do_GSEA_KEGG}
\title{Performs gene set enrichment analysis using KEGG terms}
\usage{
do_GSEA_KEGG(
  proteins,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 120,
  dataset = dataset,
  view = T,
  return.all = F,
  add.info = F
)
}
\arguments{
\item{proteins}{numeric vector of proteins scores}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction), "hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}

\item{dataset}{dataset}

\item{view}{view results}

\item{return.all}{return enrichResult object; useful for further analysis of enrichment results}

\item{add.info}{add additional information to the results data frame}
}
\value{

}
\description{
Performs gene set enrichment analysis using KEGG terms
}
