% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust.R
\name{do_hclust}
\alias{do_hclust}
\title{Performs hierarchical clustering on data frame}
\usage{
do_hclust(
  data_,
  scale = T,
  distance.method = "euclidean",
  clustering.method = "complete",
  plot = T,
  input = "LFQ.intensity",
  output = "data_hclust"
)
}
\arguments{
\item{data_}{data list}

\item{scale}{Scale data (Z-score)?}

\item{distance.method}{method to calculate distance; see ?dist for options}

\item{clustering.method}{method to cluster columns and rows; see ?hclust for options}

\item{plot}{generate default plot with plot_gg_heatmap()}

\item{input}{name of input data}

\item{output}{name of output data}
}
\value{

}
\description{
Performs hierarchical clustering on data frame
}
