% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust.R
\name{do_hclust}
\alias{do_hclust}
\title{Performs hierarchical clustering on data frame}
\usage{
do_hclust(
  data_,
  group.column = "groups",
  grouping.function = mean,
  clustering.method
)
}
\arguments{
\item{data}{data}
}
\value{

}
\description{
Performs hierarchical clustering on data frame
}
