% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_hclust_fun_enrich.R
\name{do_hclust_fun_enrich}
\alias{do_hclust_fun_enrich}
\title{Title}
\usage{
do_hclust_fun_enrich(data_, max.k = 2, input = "dend_x")
}
\arguments{
\item{data_}{}
}
\value{

}
\description{
Title
}
