% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ORA.R
\name{do_ORA}
\alias{do_ORA}
\title{Performs overexpression enrichment on logical protein vector with specified Ontology}
\usage{
do_ORA(proteins, database, algorithm, threshold, add.info = F)
}
\arguments{
\item{proteins}{character vector}

\item{database}{database to use}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{threshold}{p-value/confidence threshold to exclude terms}

\item{add.info}{Add additional information (takes longer)}
}
\value{

}
\description{
Performs overexpression enrichment on logical protein vector with specified Ontology
}
