% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ORA_KEGG.R
\name{do_ORA_KEGG}
\alias{do_ORA_KEGG}
\title{Performs over-representation analysis using KEGG annotations}
\usage{
do_ORA_KEGG(
  proteins,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  dataset,
  view = F
)
}
\arguments{
\item{proteins}{numeric/logical vector of proteins indicating group}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction), "hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}

\item{dataset}{dataset}

\item{view}{view results}
}
\description{
Performs over-representation analysis using KEGG annotations
}
