% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dend_.R
\name{enrich_dend_}
\alias{enrich_dend_}
\title{Calculates enrichment for each cluster level}
\usage{
enrich_dend_(
  cor_list,
  proteins,
  n = 10,
  inverse = F,
  name,
  cut = T,
  plot.lines = F,
  plot.matrix = T,
  labels,
  fun.enrich = F,
  database = "GO",
  algorithm = "weight01",
  add.info = F
)
}
\arguments{
\item{cor_list}{cor_list object}

\item{proteins}{annotated proteins or protein score vector}

\item{n}{max number of clusters}

\item{inverse}{inverse enrichment (works for both Fisher and KS test)}

\item{name}{name}

\item{cut}{cut dendrogram}

\item{plot.lines}{Plot lines plot}

\item{plot.matrix}{Plot correlation matrix}

\item{labels}{proteins to label on y-axis}

\item{fun.enrich}{Do functional enrichment}

\item{database}{databases to use for functional enrichment}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{add.info}{adds protein information to fun enrich tables}
}
\description{
Calculates enrichment for each cluster level
}
