% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dend_.R
\name{enrich_dend_}
\alias{enrich_dend_}
\title{Calculates enrichment for each cluster level}
\usage{
enrich_dend_(
  cor_list,
  proteins,
  n = 10,
  inverse = F,
  name,
  cut = T,
  return = F,
  plot = T,
  labels,
  fun.enrich = T,
  databases = "GO",
  algorithm = "weight01"
)
}
\arguments{
\item{cor_list}{cor_list object}

\item{proteins}{annotated proteins or protein score vector}

\item{n}{max number of clusters}

\item{inverse}{inverse enrichment (works for both Fisher and KS test)}

\item{name}{name}

\item{cut}{cut dendrogram}

\item{return}{return cor_list object}

\item{plot}{Plot results?}

\item{labels}{proteins to label on y-axis}

\item{fun.enrich}{Do functional enrichment}

\item{databases}{databases to use for functional enrichment}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}
}
\value{

}
\description{
Calculates enrichment for each cluster level
}
