% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_fisher_GO.R
\name{enrichment_fisher_GO}
\alias{enrichment_fisher_GO}
\title{Performs overexpression enrichment on logical protein vector with specified GO ontology}
\usage{
enrichment_fisher_GO(proteins, ontology)
}
\arguments{
\item{proteins}{character vector}

\item{ontology}{GO ontology}
}
\value{

}
\description{
Performs overexpression enrichment on logical protein vector with specified GO ontology
}
