% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data_var.R
\name{eval_data_var}
\alias{eval_data_var}
\title{Evaluates data column-wise}
\usage{
eval_data_var(
  data,
  expr,
  variables = "default",
  observations = "default",
  observations.set,
  data.name,
  dataset,
  view = F,
  save = F,
  name,
  set.default = F
)
}
\arguments{
\item{data}{data}

\item{expr}{function(x) how columns should be evaluated}

\item{variables}{variables}

\item{observations}{observations}

\item{observations.set}{set of observations}

\item{data.name}{data name}

\item{dataset}{dataset}

\item{view}{View results?}

\item{save}{Save results?}

\item{name}{name of new data}

\item{set.default}{should new data be set as default}
}
\value{

}
\description{
Evaluates data column-wise
}
