% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_DT_fun_enrich.R
\name{export_DT_fun_enrich}
\alias{export_DT_fun_enrich}
\title{Formats table for nice output (R Markdown)}
\usage{
export_DT_fun_enrich(data)
}
\arguments{
\item{data}{data from fun_enrich (enrichResult)}
}
\value{

}
\description{
Formats table for nice output (R Markdown)
}
