% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_.R
\name{expr_}
\alias{expr_}
\title{Title}
\usage{
expr_(data, groups, scale = T, combine = "mean", observations.set, dataset)
}
\arguments{
\item{data}{}

\item{groups}{}

\item{scale}{}

\item{combine}{}

\item{observations.set}{}

\item{dataset}{}
}
\value{

}
\description{
Title
}
