% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich.R
\name{fun_enrich}
\alias{fun_enrich}
\title{Performs appropriate enrichment on given sets of proteins or values}
\usage{
fun_enrich(
  proteins,
  background = NULL,
  databases = "GO",
  algorithm = "weight01",
  view = T,
  return = F,
  save = F
)
}
\arguments{
\item{proteins}{list of proteins/scores/protein groups}

\item{background}{(optional) background genes}

\item{databases}{databases to use}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}

\item{view}{View results?}

\item{return}{Return results?}

\item{save}{Save results?}
}
\value{

}
\description{
Performs appropriate enrichment on given sets of proteins or values
}
