% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich.R
\name{fun_enrich}
\alias{fun_enrich}
\title{Performs appropriate enrichment on given sets of proteins or values}
\usage{
fun_enrich(
  proteins,
  background = NULL,
  database = "GO",
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  inverse = F,
  dataset,
  view = T
)
}
\arguments{
\item{proteins}{list of proteins/scores/protein groups}

\item{background}{(optional) background genes}

\item{database}{database(s) to use}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction),
"hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}

\item{inverse}{scores: enriches for higher scores if TRUE}

\item{dataset}{dataset}

\item{view}{View results}

\item{threshold}{p-value/confidence threshold to exclude terms}
}
\description{
Performs appropriate enrichment on given sets of proteins or values
}
