% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich_.R
\name{fun_enrich_}
\alias{fun_enrich_}
\title{Functional enrichment of cor_list object}
\usage{
fun_enrich_(cor_list, name, k, l, databases = "CC", algorithm = "weight01")
}
\arguments{
\item{cor_list}{cor_list object}

\item{name}{(optional) name of enrichment entry}

\item{k}{(optional) number of clusters}

\item{l}{(optional) number of cluster to analyse specifically}

\item{databases}{databases for which to perform enrichment}

\item{algorithm}{algorithm to use ("classic", "elim", "weight", "weight01")}
}
\value{

}
\description{
Either name or k must be given. Number of clusters will be analysed as groups if no l is specified.
}
