% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich_t_test.R
\name{fun_enrich_t.test}
\alias{fun_enrich_t.test}
\title{Functional enrichment of t.test results}
\usage{
fun_enrich_t.test(
  data_,
  ...,
  view = F,
  dataset,
  input = "data_t.test",
  output = "fun_enrich_t.test"
)
}
\arguments{
\item{data_}{list or tibble}

\item{...}{additional arguments to fun_enrich}

\item{view}{view results}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Functional enrichment of t.test results
}
