% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Assemble data from dataset}
\usage{
get_data(variables, observations, observations.set, name, type, dataset)
}
\arguments{
\item{variables}{selected variables}

\item{observations}{selected observations}

\item{observations.set}{set of observations}

\item{name}{specific name of data type}

\item{type}{data type}

\item{dataset}{dataset name or number}
}
\value{

}
\description{
Assemble data from dataset
}
