% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Assemble data from dataset}
\usage{
get_data(
  name,
  variables = "default",
  observations = "default",
  observations.set,
  type,
  dataset
)
}
\arguments{
\item{name}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{observations.set}{set of observations}

\item{type}{data type}

\item{dataset}{dataset name or number}
}
\value{

}
\description{
Assemble data from dataset
}
