% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Assemble data from dataset}
\usage{
get_data(
  data.name,
  variables = "default",
  observations = "default",
  observations.set,
  dataset,
  tidy = T
)
}
\arguments{
\item{data.name}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{observations.set}{set of observations}

\item{dataset}{dataset name or number}

\item{tidy}{return tidy data frame or matrix}
}
\value{

}
\description{
Assemble data from dataset
}
