% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Assemble data from dataset}
\usage{
get_data(name, type, dataset, variables, observations.set, observations)
}
\arguments{
\item{name}{specific name of data type}

\item{type}{data type}

\item{dataset}{dataset name or number}

\item{variables}{selected variables}

\item{observations.set}{set of observations}

\item{observations}{selected observations}
}
\value{

}
\description{
Assemble data from dataset
}
