% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Assemble data from dataset and return in list}
\usage{
get_data(
  which,
  variables = "default",
  observations = "default",
  output,
  output.type = "list",
  observations.set,
  dataset
)
}
\arguments{
\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{output}{name to save in list}

\item{output.type}{output type ("list" (default), "tibble", "data.frame", "matrix")}

\item{observations.set}{set of observations}

\item{dataset}{dataset name or number}
}
\value{

}
\description{
Assemble data from dataset and return in list
}
