% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_.R
\name{get_data_}
\alias{get_data_}
\title{Assemble data from dataset and return in list}
\usage{
get_data_(
  data.name,
  variables = "default",
  observations = "default",
  observations.set,
  dataset,
  tidy = T,
  name = "raw_data"
)
}
\arguments{
\item{data.name}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{observations.set}{set of observations}

\item{dataset}{dataset name or number}

\item{tidy}{add data as tidy data frame or matrix}

\item{name}{name to save in list}
}
\value{

}
\description{
Assemble data from dataset and return in list
}
