% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_names.R
\name{get_data_names}
\alias{get_data_names}
\title{Prints or returns data names}
\usage{
get_data_names(dataset, print = T)
}
\arguments{
\item{dataset}{dataset}

\item{print}{print}
}
\value{

}
\description{
Prints or returns data names
}
