% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Returns groups vector}
\usage{
get_groups(
  observations,
  groups,
  control,
  observations.set,
  dataset,
  control.first = F
)
}
\arguments{
\item{observations}{observations}

\item{groups}{groups}

\item{control}{(optional) which group is control}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{control.first}{(optional) should first group be set as control}
}
\value{

}
\description{
Returns groups vector
}
