% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Return variables data}
\usage{
get_variables(variables, dataset)
}
\arguments{
\item{variables}{vector of variables}

\item{dataset}{dataset}
}
\value{

}
\description{
Return variables data
}
