% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_data.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(variables, which, names = "variables", dataset)
}
\arguments{
\item{variables}{(optional) vector of variables or expression}

\item{which}{which variables data to pull}

\item{names}{column to use as names of vector}

\item{dataset}{dataset}
}
\value{

}
\description{
Return variables data
}
