% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_data.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(variables, name, dataset)
}
\arguments{
\item{variables}{vector of variables}

\item{name}{name of data}

\item{dataset}{dataset}
}
\value{

}
\description{
Return variables data
}
