% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_variables.R
\name{identify_variables}
\alias{identify_variables}
\title{Finds feasible identifier columns}
\usage{
identify_variables(x, sep, identifier)
}
\arguments{
\item{x}{data frame}

\item{sep}{separator}

\item{identifier}{vector of identifier columns}
}
\value{

}
\description{
Finds feasible identifier columns
}
