% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_datasets.R
\name{import_datasets}
\alias{import_datasets}
\title{Import files and stores in a list; changes wd if files come from new wd}
\usage{
import_datasets(
  files,
  change.wd = F,
  prepare,
  species,
  data.origin,
  load.UniProt.ws = F
)
}
\arguments{
\item{files}{files to be imported}

\item{change.wd}{Should wd be changed by default}

\item{prepare}{Should imported files be forwarded to prepare_datasets}

\item{species}{should taxonomy be determined}

\item{data.origin}{Identify origin software of data}

\item{load.UniProt.ws}{Should a UniProt database be downloaded}
}
\value{

}
\description{
Import files and stores in a list; changes wd if files come from new wd
}
