% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta.R
\name{import_fasta}
\alias{import_fasta}
\title{Imports fasta file and saves proteome}
\usage{
import_fasta(file, name, add.proteome = T, extract.all = T, replace = F)
}
\arguments{
\item{file}{fasta file path}

\item{name}{(optional) name to save information under}

\item{add.proteome}{add all accessions of proteme}

\item{extract.all}{extract all information from fasta headers}

\item{replace}{replace existing databases}
}
\value{

}
\description{
Imports fasta file and saves proteome
}
