% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_files.R
\name{import_files}
\alias{import_files}
\title{Imports any file type using the file extension and returns list}
\usage{
import_files(files, return = F)
}
\arguments{
\item{files}{file paths}

\item{return}{return imported files}
}
\value{

}
\description{
Imports any file type using the file extension and returns list
}
