% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_norm.R
\name{impute_norm}
\alias{impute_norm}
\title{Imputes values based on normal deviation}
\usage{
impute_norm(data, shift = 1.8, width = 0.3, seed = 123)
}
\arguments{
\item{data}{data}

\item{shift}{shift in standard deviations}

\item{width}{width in standard deviations}

\item{seed}{seed}
}
\value{

}
\description{
Imputes values based on normal deviation
}
