% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_groups_.R
\name{include_groups_}
\alias{include_groups_}
\title{Adds group vector to data frame}
\usage{
include_groups_(
  analysis_list,
  groups,
  control,
  observations.set,
  dataset,
  target.data = "raw_data"
)
}
\arguments{
\item{analysis_list}{analysis list}

\item{groups}{groups}

\item{control}{(optional) which group is control}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{target.data}{data frame to be modified}
}
\value{

}
\description{
Adds group vector to data frame
}
