% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_observations_data.R
\name{include_observations_data}
\alias{include_observations_data}
\title{Adds observations data to data frame}
\usage{
include_observations_data(data, name, observations.set, dataset)
}
\arguments{
\item{data}{observations data}

\item{name}{name of data}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}
}
\value{

}
\description{
Adds observations data to data frame
}
