% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_observations_data_.R
\name{include_observations_data_}
\alias{include_observations_data_}
\title{Adds group vector to data frame}
\usage{
include_observations_data_(
  analysis_list,
  name,
  observations.set,
  dataset,
  target.data = "raw_data"
)
}
\arguments{
\item{name}{observations data}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{target.data}{data frame to be modified}

\item{data}{data frame}
}
\value{

}
\description{
Adds group vector to data frame
}
