% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_first.R
\name{keep_first}
\alias{keep_first}
\title{Performs strsplit and keeps first element of each string}
\usage{
keep_first(x, split, as_numeric = T)
}
\arguments{
\item{x}{vector}

\item{split}{separator}

\item{as_numeric}{try to convert to numeric}
}
\value{

}
\description{
Performs strsplit and keeps first element of each string
}
