% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_firstn.R
\name{keep_firstn}
\alias{keep_firstn}
\title{Removes multiple entries separated by a character}
\usage{
keep_firstn(x, n, sep)
}
\arguments{
\item{x}{vector}

\item{n}{number of elements to keep from beginning}

\item{sep}{separator}
}
\value{

}
\description{
Removes multiple entries separated by a character
}
