% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dataset.R
\name{new_dataset}
\alias{new_dataset}
\title{Sorts raw data into data list with attributes}
\usage{
new_dataset(import, name, data.type, identifier, data.columns)
}
\arguments{
\item{import}{imported raw data frame}

\item{name}{name of dataset}

\item{data.type}{output type from MaxQuant}

\item{identifier}{(optional) specific vector of identifier column/s}

\item{data.columns}{(optional) types of data to extract; uses defaults if not specified}
}
\value{

}
\description{
Sorts raw data into data list with attributes
}
