% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_imports_list.R
\name{new_imports_list}
\alias{new_imports_list}
\title{Creates imports list}
\usage{
new_imports_list(replace = F, return = T)
}
\arguments{
\item{replace}{Should imports list be replaced if already existing}

\item{return}{Should indication whether new imports list was created be returned}
}
\value{

}
\description{
Creates imports list
}
