% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_bar_anova_eval.R
\name{plot_gg_bar_anova_eval}
\alias{plot_gg_bar_anova_eval}
\title{Plots stacked bar plot to indicate significantly changed proteins from eval_anova output}
\usage{
plot_gg_bar_anova_eval(
  data,
  columns,
  labels,
  x.axis.title = "",
  y.axis.title = "Proteins",
  bar.width = 0.9,
  y.axis.breaks,
  aspect.ratio = 0.8,
  legend.position = "right",
  custom.theme = theme_hjv_half_open,
  ...
)
}
\arguments{
\item{data}{data}

\item{columns}{columns to use for plot}

\item{labels}{x-axis labels for columns}

\item{x.axis.title}{title of x-axis}

\item{y.axis.title}{title of y-axis}

\item{y.axis.breaks}{break size between ticks of y-axis}

\item{aspect.ratio}{y/x ratio}

\item{legend.position}{where should the legend be printed}

\item{custom.theme}{theme}

\item{...}{}
}
\value{

}
\description{
Plots stacked bar plot to indicate significantly changed proteins from eval_anova output
}
