% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_count_id_quant.R
\name{plot_gg_count_id_quant}
\alias{plot_gg_count_id_quant}
\title{Barplot of identified/quantified proteins per sample}
\usage{
plot_gg_count_id_quant(
  data_,
  labels,
  order.by,
  color.by = "All",
  color = list(all = c("#6BAED6", "#2171B5")),
  custom.theme = theme_phosprot_half_open,
  xlab = NULL,
  ylab = "Proteins",
  limit.y.top,
  aspect.ratio = 0.66,
  dataset,
  input = "data_count_id_quant",
  output = "plot_count_id_quant",
  View = T
)
}
\arguments{
\item{data_}{data_ list}

\item{labels}{column name for labels}

\item{order.by}{order.by factor levels of specified column}

\item{color.by}{column to use for coloring (default is "All" - column will
added automatically)}

\item{color}{list containing vectors of length two (first color id, second
quant) for each group}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{limit.y.top}{upper limit of y-axis}

\item{aspect.ratio}{y/x ratio of plot axis}

\item{input}{name of id_quant data}

\item{output}{name of plot data}

\item{View}{print plot}
}
\description{
Barplot of identified/quantified proteins per sample
}
