% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_heatmap.R
\name{plot_gg_heatmap}
\alias{plot_gg_heatmap}
\title{Plots heatmap with dendrograms}
\usage{
plot_gg_heatmap(
  data_,
  transpose = F,
  protein.label.FUN = p2g,
  observations.labels,
  dend.x.height = 0.05,
  dend.y.width = 0.2,
  labels.r = 0.3,
  x.labels.angle = 45,
  export = F,
  height = 6,
  ratio = 3,
  input = "data_hclust",
  file = "heatmap.pdf"
)
}
\arguments{
\item{data_}{data}

\item{transpose}{should data be transformed}

\item{dend.x.height}{relative height of x-axis dendrogram}

\item{dend.y.width}{relative width of y-axis dendrogram}

\item{labels.r}{width of labels (must be optimized)}

\item{x.labels.angle}{angle of x-axis labels}

\item{export}{export plot as pdf}

\item{height}{plot height in inch for export}

\item{ratio}{y/x ratio of heat map}

\item{file}{}

\item{label.proteins}{should proteins be labeled on the y-axis?}
}
\value{

}
\description{
Plots heatmap with dendrograms
}
