% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_heatmap.R
\name{plot_gg_heatmap}
\alias{plot_gg_heatmap}
\title{Plots heatmap with dendrograms}
\usage{
plot_gg_heatmap(
  data_,
  transpose = F,
  protein.label.FUN = p2g,
  observation.labels,
  labels_x = T,
  labels_y = F,
  dend_x = T,
  dend_y = T,
  annot_layer_x,
  annot_layer_y,
  legends,
  ratio = 3,
  rel_dend_x = 0.1,
  rel_dend_y = 0.2,
  rel_labels_x = 0.2,
  rel_labels_y = 0.2,
  rel_annot_layer_x = 0.05,
  rel_annot_layer_y = 0.1,
  rel_legends = 0.25,
  rel_legends_space = 0.05,
  axis.text.x.angle = 45,
  annotation.colors = list(),
  heatmap.legend.title = "",
  export = F,
  height = 6,
  dataset,
  print = T,
  input = "data_hclust",
  output = "plot_hclust",
  file = "heatmap.pdf"
)
}
\arguments{
\item{data_}{data}

\item{transpose}{should data be transformed}

\item{protein.label.FUN}{function for protein labels}

\item{observation.labels}{column for observation labels}

\item{labels_x}{(logical) plot x-axis labels}

\item{labels_y}{(logical) plot y-axis labels}

\item{dend_x}{(logical) plot x-axis dendrogram}

\item{dend_y}{(logical) plot y-axis dendrogram}

\item{annot_layer_x}{(character) data columns to be included as annotation layers on x-axis}

\item{annot_layer_y}{(character) data columns to be included as annotation layers on y-axis}

\item{legends}{name of plot elements for which a legend should be included (full names must be used, e.g. c("annot_layer_x_groups", "heatmap")}

\item{ratio}{y/x ratio of heat map}

\item{rel_dend_x}{relative height of x-axis dendrogram}

\item{rel_dend_y}{relative width of y-axis dendrogram}

\item{rel_labels_x}{relative height of x-axis labels}

\item{rel_labels_y}{relative width of y-axis labels}

\item{rel_annot_layer_x}{relative height/s of x-axis annotation layers (either one number or vector with number for each)}

\item{rel_annot_layer_y}{relative width/s of y-axis annotation layers (either one number or vector with number for each)}

\item{rel_legends}{relative width of legends layer to complete heatmap}

\item{rel_legends_space}{realtive space between plot and legends}

\item{axis.text.x.angle}{angle of x-axis labels}

\item{annotation.colors}{(optional) named list containing color vector for annotation layers}

\item{heatmap.legend.title}{title of heatmap legend}

\item{export}{export plot as pdf}

\item{height}{plot height in inch for export}

\item{dataset}{dataset}

\item{print}{print plot}

\item{input}{input data name}

\item{output}{output data name}

\item{file}{file name}
}
\value{

}
\description{
Plots heatmap with dendrograms
}
