% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_heatmap_protein_overlap.R
\name{plot_gg_heatmap_protein_overlap}
\alias{plot_gg_heatmap_protein_overlap}
\title{Compares log2 fold-changes between two comparisons from an anova_eval data list}
\usage{
plot_gg_heatmap_protein_overlap(
  data_,
  x,
  y,
  font_size = 6,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14,
  add.text = T,
  max.terms = 5,
  custom.theme = theme_hjv_overlap_heatmap
)
}
\arguments{
\item{data_}{data list from anova_eval}

\item{x}{comparison on x-axis}

\item{y}{comparison on y-axis}

\item{font_size}{Overall font size.}

\item{rel_small}{Relative size of small text (e.g., axis tick labels)}

\item{rel_tiny}{Relative size of tiny text (e.g., caption)}

\item{rel_large}{Relative size of large text (e.g., title)}

\item{add.text}{add terms in boxes}

\item{max.terms}{maximum number of terms per box}

\item{custom.theme}{theme to use for plot}
}
\value{

}
\description{
Compares log2 fold-changes between two comparisons from an anova_eval data list
}
