% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_volcano.R
\name{plot_gg_volcano}
\alias{plot_gg_volcano}
\title{Plots a volcano plot with ggplot2}
\usage{
plot_gg_volcano(
  data,
  p.value.cutoff = 0.05,
  pos.log2fc.cutoff = 0,
  neg.log2fc.cutoff = 0,
  highlight.variables,
  highlight.color = "black",
  x.axis.title = "log2 fold-change",
  y.axis.title = "-log10(p-value)",
  text.size = 6,
  text.color = "black",
  point.size = 2,
  point.alpha = 0.8,
  highlight.point.size = 3,
  highlight.point.alpha = 0.8,
  x.axis.breaks = 1,
  y.axis.breaks = 1,
  axis.line.size = 0.5,
  axis.color = "black",
  axis.ticks.size = 0.3,
  axis.title.size = 8,
  axis.text.size = 6
)
}
\arguments{
\item{data}{data}

\item{p.value.cutoff}{p-value limit for coloring}

\item{pos.log2fc.cutoff}{positive log2 fold-change limit for coloring}

\item{neg.log2fc.cutoff}{negative log2 fold-change limit for coloring}

\item{highlight.variables}{variables to highlight by point.size}

\item{x.axis.title}{title of x-axis}

\item{y.axis.title}{title of y-axis}

\item{text.size}{size of text in points (5-8)}

\item{text.color}{color of text}

\item{point.size}{point size (0.5-2)}

\item{point.alpha}{transparency (0-1)}

\item{highlight.point.size}{size of point of highlighted variables}

\item{highlight.point.alpha}{transparency of points to be highlighted}

\item{x.axis.breaks}{break size between ticks of x-axis}

\item{y.axis.breaks}{break size between ticks of y-axis}

\item{axis.line.size}{width of axes lines}

\item{axis.color}{color of axes lines}

\item{axis.ticks.size}{width of axis ticks}

\item{axis.title.size}{size of axis title}

\item{axis.text.size}{size of axis labels}

\item{highligh.color}{color to use to highlight specified proteins}
}
\value{

}
\description{
Plots a volcano plot with ggplot2
}
