% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_discrete.R,
%   R/plot_heatmap_discrete_OLD.R
\name{plot_heatmap_discrete}
\alias{plot_heatmap_discrete}
\title{Plots heatmap with dendrograms}
\usage{
plot_heatmap_discrete(
  data,
  transform = T,
  clustering.method = "complete",
  order.y = T,
  order.x = T,
  colors,
  labels,
  legend.title = NULL,
  legend.text.align = 0,
  ratio = 1,
  label.fun
)

plot_heatmap_discrete(
  data,
  transform = T,
  clustering.method = "complete",
  order.y = T,
  order.x = T,
  colors,
  labels,
  legend.title = NULL,
  legend.text.align = 0,
  ratio = 1,
  label.fun
)
}
\arguments{
\item{data}{data frame with first column variables}

\item{transform}{should data be transformed?}

\item{clustering.method}{argument for hclust}

\item{order.y}{either T/F if axis should be reordered by dendrogram, or vector containing all axis labels in order}

\item{order.x}{either T/F if axis should be reordered by dendrogram, or vector containing all axis labels in order}

\item{colors}{named vector of colors with names as corresponding values}

\item{labels}{named vector of labels with names as corresponding values}

\item{legend.title}{title of legend}

\item{legend.text.align}{0 = left, 1 = right}

\item{ratio}{x-y-ratio of heatmap tiles}

\item{label.fun}{function to generate labels with}

\item{data_}{data}

\item{transpose}{should data be transformed}

\item{protein.label.FUN}{function for protein labels}

\item{observation.labels}{column for observation labels}

\item{labels_x}{(logical) plot x-axis labels}

\item{labels_y}{(logical) plot y-axis labels}

\item{dend_x}{(logical) plot x-axis dendrogram}

\item{dend_y}{(logical) plot y-axis dendrogram}

\item{annot_layer_x}{(character) data columns to be included as annotation
layers on x-axis}

\item{annot_layer_y}{(character) data columns to be included as annotation
layers on y-axis}

\item{rel_dend_x}{relative height of x-axis dendrogram}

\item{rel_dend_y}{relative width of y-axis dendrogram}

\item{rel_labels_x}{relative height of x-axis labels}

\item{rel_labels_y}{relative width of y-axis labels}

\item{rel_annot_layer_x}{relative height/s of x-axis annotation layers
(either one number or vector with number for each)}

\item{rel_annot_layer_y}{relative width/s of y-axis annotation layers
(either one number or vector with number for each)}

\item{axis.text.x.angle}{angle of x-axis labels}

\item{annotation.colors}{(optional) named list containing color vector for
annotation layers}

\item{legends}{name of plot elements for which a legend should be included
(full names must be used, e.g. c("annot_layer_x_groups", "heatmap"))}

\item{rel_legends}{relative width of legends layer to complete heatmap}

\item{rel_legends_space}{relative space between plot and legends}

\item{heatmap.legend.title}{title of heatmap legend}

\item{dataset}{dataset}

\item{file}{file name}

\item{view}{print plot}

\item{export}{export plot as pdf}

\item{height}{plot height in inch for export}

\item{input}{input data name}

\item{output}{output data name}
}
\value{



}
\description{
Plots heatmap with dendrograms

Plots heatmap from discrete values
}
