% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_dxy.R
\name{plot_heatmap_dxy}
\alias{plot_heatmap_dxy}
\title{Plots heatmap with dendrograms}
\usage{
plot_heatmap_dxy(
  data,
  dend_x,
  dend_y,
  scale = T,
  transform = T,
  clustering.method = "complete"
)
}
\arguments{
\item{data}{data}

\item{dend_x}{x-axis dendrogram}

\item{dend_y}{y-axis dendrogram}

\item{scale}{should data be scaled}

\item{transform}{should data be transformed}

\item{clustering.method}{method to cluster variables and observations}
}
\value{

}
\description{
Plots heatmap with dendrograms
}
