% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{Visualizes identification data as UpSet plot}
\usage{
plot_upset(
  data_,
  order,
  nintersects = 10,
  transpose,
  from.row.names,
  view = T,
  input,
  output = "plot_upset"
)
}
\arguments{
\item{data_}{data_ list}

\item{order}{order of samples}

\item{nintersects}{number of bars representing intersects}

\item{transpose}{transpose data frame}

\item{from.row.names}{row names column}

\item{view}{view plot}

\item{input}{name of input data}

\item{output}{name of output plot}
}
\description{
Visualizes identification data as UpSet plot
}
