% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plots a venn diagram form given data}
\usage{
plot_venn(x, name, plot = T, export = F)
}
\arguments{
\item{x}{table or list}

\item{name}{name}

\item{plot}{plot venn diagrams to device}

\item{export}{exports plot to pdf}
}
\value{

}
\description{
Plots a venn diagram form given data
}
