% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Title}
\usage{
plot_volcano(
  x,
  y,
  markers = "default",
  size = 4,
  opacity = 0.8,
  axis.title.x = "log2 fold-change",
  axis.title.y = "-log10 p-value"
)
}
\arguments{
\item{x}{log2 fold change}

\item{y}{pvalue}

\item{markers}{specific proteins to highlight}

\item{size}{size of dots}

\item{opacity}{opacity of dots}

\item{axis.title.x}{}

\item{axis.title.y}{}
}
\value{

}
\description{
Title
}
