% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy.R
\name{plot_xy}
\alias{plot_xy}
\title{Matches names and plots scattrplot of two variables}
\usage{
plot_xy(x, y, axis.title.x = "x", axis.title.y = "y")
}
\arguments{
\item{x}{names x vector}

\item{y}{named y vector}

\item{fit}{add fit}
}
\value{

}
\description{
Matches names and plots scattrplot of two variables
}
