% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{prepare_datasets}
\alias{prepare_datasets}
\title{Prepares imported data}
\usage{
prepare_datasets(
  imports,
  data.origin,
  species,
  load.UniProt.ws = F,
  min.similarity = 8,
  min.groupsize = 6
)
}
\arguments{
\item{imports}{imported raw data}

\item{data.origin}{Identify origin software of data}

\item{species}{protein origin (T or F or UniProt taxosnomy ID)}

\item{load.UniProt.ws}{Should a UniProt database be downloaded}

\item{min.similarity}{minimum similarity oi column names}

\item{min.groupsize}{minimum number of samples}
}
\value{

}
\description{
Prepares imported data
}
