% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{prepare_datasets}
\alias{prepare_datasets}
\title{Prepares imported data}
\usage{
prepare_datasets(
  imports,
  data.origin,
  species,
  load.UniProt.ws = F,
  identifier,
  data.types
)
}
\arguments{
\item{imports}{imported raw data}

\item{data.origin}{Identify origin software of data}

\item{species}{protein origin (T or F or UniProt taxonomy ID)}

\item{load.UniProt.ws}{Should a UniProt database be downloaded}

\item{identifier}{(optional) specific vector of column/s to use as identifier}

\item{data.types}{types of data to be extracted}
}
\value{

}
\description{
Prepares imported data
}
