% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_.R
\name{scale_}
\alias{scale_}
\title{Scales data (Z-scores) from tibbles or analysis_list containing a tibble}
\usage{
scale_(analysis_list, target.data = "raw_data")
}
\arguments{
\item{analysis_list}{data list}

\item{target.data}{data to be modified}
}
\value{

}
\description{
Scales data (Z-scores) from tibbles or analysis_list containing a tibble
}
