% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_UniProt.R
\name{select_UniProt}
\alias{select_UniProt}
\title{Improved select function for UniProt databases}
\usage{
select_UniProt(x, keys, columns, keytype = "UNIPROTKB", modify = T)
}
\arguments{
\item{x}{UniProt data base}

\item{keys}{keys}

\item{columns}{data to query}

\item{keytype}{type of key}

\item{modify}{should columns be modified after query}
}
\value{

}
\description{
Improved select function for UniProt databases
}
