% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_UniProt.R
\name{select_UniProt}
\alias{select_UniProt}
\title{Improved select function for UniProt databases}
\usage{
select_UniProt(
  keys,
  columns,
  output = "vector.keep",
  keytype = "UNIPROTKB",
  x,
  modify = T,
  dataset
)
}
\arguments{
\item{keys}{keys}

\item{columns}{data to query}

\item{output}{format of output ("vector.keep" (default), "vector.na,
"vector.rm", "mapping.na", "mapping.rm", "TERM2GENE")}

\item{keytype}{type of keys}

\item{x}{UniProt data base}

\item{modify}{(recommended) Should output values be modified to to singular
entries?}

\item{dataset}{dataset}
}
\description{
Improved select function for UniProt databases
}
