% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_msigdb.R
\name{setup_msigdb}
\alias{setup_msigdb}
\title{Downloads and prepares MSigDB annotations}
\usage{
setup_msigdb(dataset, keep.loaded = F, save = T, dir = "Data/Databases/")
}
\arguments{
\item{dataset}{dataset}
}
\value{

}
\description{
Downloads and prepares MSigDB annotations
}
