% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_STRINGdb.R
\name{setup_STRINGdb}
\alias{setup_STRINGdb}
\title{Sets up the STRING database}
\usage{
setup_STRINGdb(
  version = "11",
  score_threshold = 400,
  dataset,
  input_directory = getwd()
)
}
\arguments{
\item{version}{STRING database version}

\item{score_threshold}{minimum score for interaction}

\item{dataset}{dataset}
}
\description{
Sets up the STRING database
}
