% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_taxonomy_information.R
\name{setup_taxonomy_information}
\alias{setup_taxonomy_information}
\title{Sets up annotaion database for biological Id translation and functional enrichment}
\usage{
setup_taxonomy_information(dataset, taxId)
}
\arguments{
\item{dataset}{dataset}

\item{taxId}{taxonomy Id}
}
\value{

}
\description{
Sets up annotaion database for biological Id translation and functional enrichment
}
