% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant.R
\name{significant}
\alias{significant}
\title{Returns logical vector indication significant p-values}
\usage{
significant(p.values, method = "threshold", threshold = 0.05, na = F)
}
\arguments{
\item{p.values}{named vector of p-values}

\item{method}{method to determine significant p-values ("threshold", "bhc" (Benjamini-Hochberg-Correction))}

\item{threshold}{threshold for p.values and FDR for Benjamini-Hochberg-Correction}
}
\value{

}
\description{
Returns logical vector indication significant p-values
}
