% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble2data.frame.R
\name{tibble2data.frame}
\alias{tibble2data.frame}
\title{Transforms tibble to data frame}
\usage{
tibble2data.frame(tibble, row.names)
}
\arguments{
\item{tibble}{tibble}

\item{row.names}{column to use for row names}
}
\value{

}
\description{
Transforms tibble to data frame
}
