% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_helpers.R
\name{tibble2data.frame}
\alias{tibble2data.frame}
\title{Transforms tibble to data frame}
\usage{
tibble2data.frame(tibble, from.row.names)
}
\arguments{
\item{tibble}{tibble}

\item{from.row.names}{column to use as new row names}
}
\value{

}
\description{
Transforms tibble to data frame
}
