% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_helpers.R
\name{tibble2matrix}
\alias{tibble2matrix}
\title{Transforms tibble to matrix}
\usage{
tibble2matrix(tibble, from.row.names)
}
\arguments{
\item{tibble}{tibble}

\item{from.row.names}{row to use as new row names}
}
\value{

}
\description{
Transforms tibble to matrix
}
