% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_names.R
\name{update_data_names}
\alias{update_data_names}
\title{Add new data name}
\usage{
update_data_names(type, dataset, default.name = NA)
}
\arguments{
\item{type}{data type}

\item{dataset}{dataset}

\item{default.name}{new default name}
}
\value{

}
\description{
Add new data name
}
