% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_types.R
\name{update_data_types}
\alias{update_data_types}
\title{Adds new data type}
\usage{
update_data_types(type, dataset, set.default = F)
}
\arguments{
\item{type}{data type}

\item{dataset}{dataset}

\item{set.default}{should default be changed}
}
\value{

}
\description{
Adds new data type
}
