% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_datasets.R
\name{update_datasets}
\alias{update_datasets}
\title{Update list of datasets}
\usage{
update_datasets(default.dataset = NA)
}
\arguments{
\item{name}{name of new dataset}

\item{set.default}{set new dataset as default}
}
\value{

}
\description{
Update list of datasets
}
